Create ODF Object From Output of read_ODF in ODF package

As of August 11, 2015, ODF::read_ODF returns a list with 9 elements, one named DATA, which is a data.frame() containing the columnar data, the others being headers of various sorts. The present function constructs an oce object from such data, facilitating processing and plotting with the general oce functions. This involves storing the 8 headers verbatim in the odfHeaders in the metadata slot, and also copying some of the header information into more standard names (e.g. metadata@longitude is a copy of metadata@odfHeader$EVENT_HEADER$INITIAL_LATITUDE). As for the DATA, they are stored in the data slot, after renaming from ODF to oce convention using ODFNames2oceNames().

Usage

ODF2oce(ODF, coerce = TRUE, debug = getOption("oceDebug"))

Arguments

ODF

A list as returned by read_ODF in the ODF package

coerce

A logical value indicating whether to coerce the return value to an appropriate object type, if possible.

debug

a flag that turns on debugging. Set to 1 to get a moderate amount of debugging information, or to 2 to get more.

Value

An oce object, possibly coerced to a subtype.

Caution

This function may change as the ODF package changes. Since ODF has not been released yet, this should not affect any users except those involved in the development of oce and ODF.

See also

Author

Dan Kelley