Map AD2CP ID Code to oce Name

As explained in Nortek (2022, section 6.1, page 80), AD2CP files use a hexadecimal (in R, "raw") code to indicate the nature of each data chunk, and read.adp.ad2cp() uses the present function as it analyses AD2CP files.

Usage

ad2cpCodeToName(code = NULL, prefix = TRUE)

Arguments

code

a raw (or corresponding integer) vector indicating the IDs of interest, or NULL to get a summary of possible values.

prefix

logical value indicating whether to show the raw value as a prefix (e.g. "0x1c=echosounder" as opposed to "echosounder").

Value

An indication of the mapping. If code is NULL, this is a data frame. Otherwise, it is a character vector with the relevant mappings, with the raw form of the code linked with the name, as in the example.

Details

The mapping from code (hex or decimal) to oce name is as follows.

code (raw)	code (integer)	oce name

———-	————–	—————–

0x15	21	burst

0x16	22	average

0x17	23	bottomTrack

0x18	24	interleavedBurst

0x1a	26	burstAltimeterRaw

0x1b	27	DVLBottomTrack

0x1c	28	echosounder

0x1d	29	DVLWaterTrack

0x1e	30	altimeter

0x1f	31	averageAltimeter

0x23	35	echosounderRaw

0xa0	160	text

References

Nortek AS. “Signature Integration 55|250|500|1000kHz.” Nortek AS, March 31, 2022.

See also

Author

Dan Kelley

Examples

stopifnot(ad2cpCodeToName(0x15) == "0x15=burst")