Convert Raw to Numeric Values in an adp Object

Convert variables in an adp object from raw to numeric format.

Usage

adpConvertRawToNumeric(
  object = NULL,
  variables = NULL,
  debug = getOption("oceDebug")
)

Arguments

object

an adp object.

variables

variables stored in an adp object that has the same dimensional as v and is stored in a raw format.

debug

an integer specifying whether debugging information is to be printed during the processing. This is a general parameter that is used by many oce functions. Generally, setting debug=0 turns off the printing, while higher values suggest that more information be printed. If one function calls another, it usually reduces the value of debug first, so that a user can often obtain deeper debugging by specifying higher debug values.

Value

adpConvertRawToNumeric returns an adp object whose specified variables have been converted from raw to numerical format.

See also

Author

Jaimie Harbin and Dan Kelley

Examples

library(oce)
data(adp)
adp[["a"]][, , 1][, 1]
#>  [1] 99 9d 9f 87 8a 99 9c 96 91 91 92 99 94 94 96 9c 98 a7 a2 94 92 8d 90 8f 90
ADP <- adpConvertRawToNumeric(adp)
ADP[["a"]][, , 1][, 1]
#>  [1] 153 157 159 135 138 153 156 150 145 145 146 153 148 148 150 156 152 167 162
#> [20] 148 146 141 144 143 144