Coerce Data Into an argo Object

Coerce a dataset into an argo dataset. This is not the right way to read official argo datasets, which are provided in NetCDF format and may be read with read.argo().

Usage

as.argo(
  time,
  longitude,
  latitude,
  salinity,
  temperature,
  pressure,
  units = NULL,
  id,
  filename = "",
  missingValue
)

Arguments

time

a vector of POSIXct times.

longitude

a vector of longitudes.

latitude

a vector of latitudes.

salinity

a vector of salinities.

temperature

a vector of temperatures.

pressure

a vector of pressures.

units

an optional list containing units. If NULL, the default, then "degree east" is used for longitude, "degree north" for latitude, "" for salinity, "ITS-90" for temperature, and "dbar" for pressure.

id

an identifier for the argo float, typically a number, but stored within the object in a character form. (For example, the dataset retrieved with data(argo) has an id of "6900388".)

filename

a source filename, which defaults to an empty string.

missingValue

an optional missing value, indicating data values that should be taken as NA.

Value

An argo object.

See also

Author

Dan Kelley