Coerce Data Into a Rsk Object

Create a rsk object.

as.rsk(
  time,
  columns,
  filename = "",
  instrumentType = "rbr",
  serialNumber = "",
  model = "",
  sampleInterval = NA,
  debug = getOption("oceDebug")
)

Arguments

time
a vector of times for the data.

columns
a list or data frame containing the measurements at the indicated times; see “Details”.

filename
optional name of file containing the data.

instrumentType
type of instrument.

serialNumber
serial number for instrument.

model
instrument model type, e.g. "RBRduo".

sampleInterval
sampling interval. If given as NA, then this is estimated as the median difference in times.

debug
a flag that can be set to TRUE to turn on debugging.

Value

An rsk object.

Details

The contents of columns are be copied into the data slot of the returned object directly, so it is critical that the names and units correspond to those expected by other code dealing with rsk objects. If there is a conductivity, it must be called conductivity, and it must be in units of mS/cm. If there is a temperature, it must be called temperature, and it must be an in-situ value recorded in ITS-90 units. And if there is a pressure, it must be absolute pressure (sea pressure plus atmospheric pressure) and it must be named pressure. No checks are made within as.rsk on any of these rules, but if they are broken, you may expect problems with any further processing.

See also