Create an xbt object

Create an xbt object

as.xbt(
  z,
  temperature,
  longitude = NA,
  latitude = NA,
  filename = "",
  sequenceNumber = NA,
  serialNumber = ""
)

Arguments

z
numeric vector giving vertical coordinates of measurements. This is the negative of depth, i.e. z is 0 at the air-sea interface, and negative within the water column.

temperature
numeric vector giving in-situ temperatures at the z values.

longitude, latitude
location in degE and degN.

filename
character value naming source file.

sequenceNumber
numerical value of the sequence number of the XBT drop.

serialNumber
character value holding the serial number of the XBT.

Value

An xbt object.

See also