Convert adv Object from Beam Coordinates to XYZ Coordinates

Convert ADV velocity components from a beam-based coordinate system to a xyz-based coordinate system.

Usage

beamToXyzAdv(x, debug = getOption("oceDebug"))

Arguments

x

an adv object.

debug

a flag that, if non-zero, turns on debugging. Higher values yield more extensive debugging.

Details

The coordinate transformation is done using the transformation matrix contained in transformation.matrix in the metadata slot, which is normally inferred from the header in the binary file. If there is no such matrix (e.g. if the data were streamed through a data logger that did not capture the header), beamToXyzAdv the user will need to store one in x, e.g. by doing something like the following:

x[["transformation.matrix"]] <- rbind(c(11100, -5771, -5321),
                                      c( #' 291, 9716, -10002),
                                      c( 1409, 1409, 1409)) / 4096

References

https://nortek.zendesk.com/hc/en-us/articles/360029820971-How-is-a-Coordinate-transformation-done-

See also

Author

Dan Kelley