Bin-map an adp Object

Bin-map an ADP object, by interpolating velocities, backscatter amplitudes, etc., to uniform depth bins, thus compensating for the pitch and roll of the instrument. This only makes sense for ADP objects that are in beam coordinates.

Usage

binmapAdp(x, debug = getOption("oceDebug"))

Arguments

x

an adp object.

debug

an integer specifying whether debugging information is to be printed during the processing. This is a general parameter that is used by many oce functions. Generally, setting debug=0 turns off the printing, while higher values suggest that more information be printed. If one function calls another, it usually reduces the value of debug first, so that a user can often obtain deeper debugging by specifying higher debug values.

Value

An adp object.

Bugs

This only works for 4-beam RDI ADP objects.

Sample of Usage

library(oce)
file <- "~/data/archive/sleiwex/2008/moorings/m09/adp/rdi_2615/raw/adp_rdi_2615.000"
beam <- read.oce(file,
    from=as.POSIXct("2008-06-26", tz="UTC"),
    to=as.POSIXct("2008-06-26 00:10:00", tz="UTC"),
    longitude=-69.73433, latitude=47.88126)
beam2 <- binmapAdp(beam)
plot(enuToOther(toEnu(beam), heading=-31.5))
plot(enuToOther(toEnu(beam2), heading=-31.5))
plot(beam, which=5:8) # backscatter amplitude
plot(beam2, which=5:8)

References

The method was devised by Clark Richards for use in his PhD work at Department of Oceanography at Dalhousie University.

See also

Author

Dan Kelley and Clark Richards