Find Profiles Within a ctd Object Read From a RBR File

This uses information about profiles that is contained within the metadata slot of the first parameter, x, having been inserted there by read.rsk(). If x was created by reading an .rsk file with read.rsk(), and if that file contained geographical information (that is, if it had a data table named geodata) then the first longitude and latitude from each profile is stored in the metadata slot of the returned value.

Usage

ctdFindProfilesRBR(
  x,
  direction = "descending",
  arr.ind = FALSE,
  debug = getOption("oceDebug")
)

Arguments

x

either an rsk or a ctd object; in the former case, it is converted to a ctd object with as.ctd().

direction

character value, either "descending" or "ascending", indicating the sampling direction to be selected. The default, "descending", is the commonly preferred choice.

arr.ind

logical value indicating whether the array indices should be returned; the alternative is to return a vector of ctd objects.

debug

an integer specifying whether debugging information is to be printed during the processing. This is a general parameter that is used by many oce functions. Generally, setting debug=0 turns off the printing, while higher values suggest that more information be printed. If one function calls another, it usually reduces the value of debug first, so that a user can often obtain deeper debugging by specifying higher debug values.

See also

Author

Dan Kelley