Sample ctd Data, Not Trimmed of Extraneous Data

This is sample CTD profile provided for testing. It includes not just the (useful) portion of the dataset during which the instrument was being lowered, but also data from the upcast and from time spent near the surface. Spikes are also clearly evident in the pressure record. With such real-world wrinkles, this dataset provides a good example of data that need trimming with ctdTrim().

Usage

data(ctdRaw)

Details

This station was sampled by students enrolled in the Dan Kelley's Physical Oceanography class at Dalhousie University. The data were acquired near the centre of the Bedford Basin of the Halifax Harbour, during an October 2003 field trip of Dalhousie University's Oceanography 4120/5120 class. (Note that the startTime in the metadata slot was altered from 1903 to 2003, using oceEdit(). The change was done because the original time was clearly incorrect, perhaps owing to the use of software that was designed to work in the twentieth only.)

See also