Suggest a Default Flag Vector for Bad or Suspicious Data

defaultFlags tries to suggest a reasonable default flag scheme for use by handleFlags(). It does this by looking for an item named flagScheme in the metadata slot of object. If flagScheme is found, and if the scheme is recognized, then a numeric vector is returned that indicates bad or questionable data. If flagScheme$default exists, then that scheme is returned. However, if that does not exist, and if flagScheme$name is recognized, then a pre-defined (very conservative) scheme is used, as listed below.

Usage

defaultFlags(object)

Arguments

object

An oce object

Value

A vector of one or more flag values, or NULL if object metadata slot lacks a flagScheme as set by initializeFlagScheme(), or if it has a scheme that is not in the list provide in “Description”.

Details

for argo, the default is c(0,3,4,6,7,9), meaning to act upon not_assessed (0), probably_bad (3), bad (4), not_used_6 (6), not_used_7 (7) and missing (9). See Section 3.2.2 of Carval et al. (2019).

for BODC, the default is c(0,2,3,4,5,6,7,8,9), i.e. all flags except good.

for DFO, the default is c(0,2,3,4,5,8,9), i.e. all flags except appears_correct.

for WHP bottle, the default is c(1,3,4,5,6,7,8,9), i.e. all flags except no_problems_noted.

for WHP ctd, the default is c(1,3,4,5,6,7,9), i.e. all flags except acceptable.

References

Carval, Thierry, Bob Keeley, Yasushi Takatsuki, Takashi Yoshida, Stephen Loch Loch, Claudia Schmid, and Roger Goldsmith. Argo User's Manual V3.3. Ifremer, 2019. doi:10.13155/29825

See also