Extract The Start of an Oce Object

Extract The Start of an Oce Object

This function handles the following object classes directly: adp, adv, argo (selection by profile), coastline, ctd, echosounder (selection by ping), section (selection by station) and topo (selection by longitude and latitude). It does not handle amsr or landsat yet, instead issuing a warning and returning x in those cases. For all other classes, it calls head() with n as provided, for each item in the data slot, issuing a warning if that item is not a vector; the author is quite aware that this may not work well for all classes. The plan is to handle all appropriate classes by July 2018. Please contact the author if there is a class you need handled before that date.

# S3 method for oce
head(x, n = 6L, ...)

Arguments

x

an oce object.

n

Number of elements to extract, as for head().

...

ignored

See also

Author

Dan Kelley