Add a Band to a landsat Object

Add a band to a landsat object. Note that it will be stored in numeric form, not raw form, and therefore it will require much more storage than data read with read.landsat().

Usage

landsatAdd(x, data, name, debug = getOption("oceDebug"))

Arguments

x

a landsat object.

data

A matrix of data, with dimensions matching that of entries already in x.

name

The name to be used for the data, i.e. the data can later be accessed with d[[name]] where d is the name of the return value from the present function.

debug

A flag that turns on debugging. Set to 1 to get a moderate amount of debugging information, or a higher value for more debugging.

Value

A landsat object, with a new data band.

See also

Author

Dan Kelley