Convert Longitude and Latitude to UTM

Convert Longitude and Latitude to UTM

lonlat2utm(longitude, latitude, zone, km = FALSE)

Arguments

longitude
numeric vector of decimal longitude. May also be a list containing items named longitude and latitude, in which case the indicated values are used, and next argument is ignored.

latitude
numeric vector of decimal latitude (ignored if longitude is a list containing both coordinates)

zone
optional indication of UTM zone. Normally this is inferred from the longitude, but specifying it can be helpful in dealing with Landsat images, which may cross zones and which therefore are described by a single zone.

km
logical value indicating whether easting and northing are in kilometers or meters.

Value

A list containing easting, northing, zone and hemisphere.

References

https://en.wikipedia.org/wiki/Universal_Transverse_Mercator_coordinate_system, downloaded May 31, 2014.

See also

Author

Dan Kelley

Examples

# \donttest{
library(oce)
## Cape Split, in the Minas Basin of the Bay of Fundy
lonlat2utm(-64.496567, 45.334626)

#> $easting
#> [1] 382736.4
#> 
#> $northing
#> [1] 5021214
#> 
#> $zone
#> [1] 20
#> 
#> $hemisphere
#> [1] "N"
#> 

# }