Sample met Data

This is sample met object containing data for Halifax, Nova Scotia, during September of 2003 (the period during which Hurricane Juan struck the city).

Source

Environment Canada website on October 19, 2019.

Details

The data file was downloaded

metFile <- download.met(id=6358, year=2003, month=9, destdir=".", type="xml")

Note that using download.met() avoids having to navigate the the awkward Environment Canada website, but it imposes the burden of having to know the station ID number. With the data in-hand, the object was then created (and its timezone adjusted) with

met <- read.met(metFile)
met <- oceSetData(met, "time", met[["time"]]+4*3600,
                 note="add 4h to local time to get UTC time")

Historical note. The data(met) object was changed on October 19, 2019, based on the data provided by Environment Canada at that time. The previous version of data(met), created in 2017, had been based on a data format that Environment Canada no longer provided in 2019. See the notes on the type argument of read.met() for more on this shift in the Environment Canada data format.

See also