Base Class for oce Objects

This is mainly used within oce to create sub-classes, although users may wish to mek direct calls to new("oce") for their own purposes.

Slots

metadata

A list containing information about the data. The contents vary across sub-classes, e.g. an adp object has information about beam patterns, which obviously would not make sense for a ctd object In addition, all classes have items named units and flags, used to store information on the units of the data, and the data quality.

data

A list containing the data.

processingLog

A list containing time-stamped processing steps, typically stored in the object by oce functions.

See also

Examples

str(new("oce"))
#> Formal class 'oce' [package "oce"] with 3 slots
#>   ..@ metadata     :List of 2
#>   .. ..$ units: list()
#>   .. ..$ flags: list()
#>   ..@ data         : list()
#>   ..@ processingLog:List of 2
#>   .. ..$ time : POSIXct[1:1], format: "2025-01-15 15:02:29"
#>   .. ..$ value: chr "Create oce object"