Write the Data Portion of Object to a File

The output has a line containing the names of the columns in x$data, each enclosed in double quotes. After that line are lines for the data themselves. The default is to separate data items by a single space character, but this can be altered by using a sep argument in the ... list; see utils::write.table().

oce.write.table(x, file = "", ...)

Arguments

x
an oce object.

file
file name, as passed to utils::write.table(). Use "" to get a listing in the terminal window.

...
optional arguments passed to utils::write.table().

Value

The value returned by utils::write.table().

Details

This function is little more than a thin wrapper around utils::write.table(), the only difference being that row names are omitted here, making for a file format that is more conventional in Oceanography.

See also

Author

Dan Kelley