Wrapper to give normalized spectrum

This is a wrapper around the R spectrum() function, which returns spectral values that are adjusted so that the integral of those values equals the variance of the input x.

oceSpectrum(x, ...)

Arguments

x

a univariate or multivariate time series, as for spectrum().

...

extra arguments passed on to spectrum().

Value

A spectrum that has values that integrate to the variance.

See also

Author

Dan Kelley

Examples

  x <- rnorm(1e3)
  s <- spectrum(x, plot=FALSE)
  ss <- oce.spectrum(x, plot=FALSE)
  cat("variance of x=", var(x), "\n")
#> variance of x= 1.004501 
  cat("integral of     spectrum=", sum(s$spec)*diff(s$freq[1:2]), "\n")
#> integral of     spectrum= 0.5077857 
  cat("integral of oce.spectrum=", sum(ss$spec)*diff(ss$freq[1:2]), "\n")
#> integral of oce.spectrum= 1.004501