Plot an ladp Object

Uses plotProfile() to create panels of depth variation of easterly and northerly velocity components.

Usage

# S4 method for class 'ladp'
plot(x, which = c("u", "v"), ...)

Arguments

x

an ladp object.

which

a character vector storing names of items to be plotted.

...

Other arguments, passed to plotting functions.

See also

Author

Dan Kelley