Get the present time, in a stated timezone

Get the present time, in a stated timezone

presentTime(tz = "UTC")

Arguments

tz

String indicating the desired timezone. The default is to use UTC, which is used very commonly in oceanographic work. To get the local time, use tz="" or tz=NULL,

Value

A POSIXct()-style object holding the present time, in the indicated timezone.

Examples

presentTime() # UTC
#> [1] "2022-03-17 11:05:27 UTC"
presentTime("") # the local timezone
#> [1] "2022-03-17 08:05:27 ADT"