Append an Item to a Processing Log

Append an Item to a Processing Log

Usage

processingLogAppend(h, value = "")

Arguments

h

either the processingLog slot of an object, or an oce object from which the processingLog will be extracted

value

A string indicating the text of the log entry.

Value

An list() containing items named time and value, i.e. the times of entries and the text notations of those entries..

See also