Read an amsr File

Read a compressed amsr file, generating an amsr object. Note that only compressed files are read in this version.

read.amsr(file, debug = getOption("oceDebug"))

Arguments

file
String indicating the name of a compressed file. See “File sources”.

debug
A debugging flag, integer.

File sources

AMSR files are provided at the FTP site ftp://ftp.ssmi.com/amsr2/bmaps_v07.2/ and login as "guest", enter a year-based directory (e.g. y2016 for the year 2016), then enter a month-based directory (e.g. m08 for August, the 8th month), and then download a file for the present date, e.g. f34_20160803v7.2.gz for August 3rd, 2016. Do not uncompress this file, since read.amsr can only read uncompressed files. If read.amsr reports an error on the number of chunks, try downloading a similarly-named file (e.g. in the present example, read.amsr("f34_20160803v7.2_d3d.gz") will report an error about inability to read a 6-chunk file, but read.amsr("f34_20160803v7.2.gz") will work properly.

See also