Read an argo File in Copernicus Format

This function was added to read a particular file downloaded from the Fleet Monitoring website (Reference 1). The format was inferred through examination of the file and a brief study of a document (Reference 2) that describes the format. Not all fields are read by this function; see Reference 3 for a full list and note that the author would be happy to add new entries (but not to spend hours entering then all).

Usage

read.argo.copernicus(
  file,
  encoding = NA,
  debug = getOption("oceDebug"),
  processingLog,
  ...
)

Arguments

file

A character string giving the name of the file to load.

encoding

ignored.

debug

A flag that turns on debugging. Set to 1 to get a moderate amount of debugging information, or 0 (the default) for silent operation.

processingLog

ignored.

...

ignored.

References

https://fleetmonitoring.euro-argo.eu/float/4902489

Copernicus Marine In Situ Tac Data Management Team. Copernicus Marine In Situ NetCDF Format Manual (version V1.43). Pdf. Copernicus in situ TAC, 2021. https://doi.org/10.13155/59938 (link checked 2022-04-11).

Variable names are provided in files at https://doi.org/10.13155/53381 (link checked 2022-04-12)

See also

Author

Dan Kelley