Read a Coastline File in Shapefile Format

Read coastline data stored in the shapefile format (see reference 1).

read.coastline.shapefile(
  file,
  lonlim = c(-180, 180),
  latlim = c(-90, 90),
  debug = getOption("oceDebug"),
  monitor = FALSE,
  processingLog
)

Arguments

file

name of file containing coastline data (a file ending in .shp) or a zipfile that contains such a file, with a corresponding name. The second scheme is useful for files downloaded from the NaturalEarth website (see reference 2).

lonlim, latlim

numerical vectors specifying the west and east edges (and south and north edges) of a focus window. Coastline polygons that do not intersect the defined box are skipped, which can be useful in narrowing high-resolution world-scale data to a local application.

debug

set to TRUE to print information about the header, etc.

monitor

Logical indicating whether to print an indication of progress through the file.

processingLog

if provided, the action item to be stored in the log. (Typically only provided for internal calls; the default that it provides is better for normal calls by a user.)

Value

x a coastline object.

A hack for depth contours

The following demonstrates that this code is getting close to working with depth contours. This should be handled more internally, and a new object for depth contours should be constructed, of which coastlines could be a subset.

References

The ``shapefile'' format is described in ESRI Shapefile Technical Description, March 1998, available at https://www.esri.com/content/dam/esrisites/sitecore-archive/Files/Pdfs/library/whitepapers/pdfs/shapefile.pdf (last checked 2021-03-24).

The NaturalEarth website https://www.naturalearthdata.com/downloads/ provides coastline datasets in three resolutions, along with similar files lakes and rivers, for borders, etc. It is highly recommended.

See also

Author

Dan Kelley