Rename variables according to a specified dictionary

The is a provisional function, added in September 2024, and likely to change through the end of that year.

Usage

rename(x, dictionary = "ioos.csv", debug = 0)

Arguments

x

an oce object.

dictionary

either a string or a data frame. If a string, then it is either the name of a built-in vocabulary (e.g. sbe; see ‘Details’) or the name of a CSV file that defines a dictionary. If it is a data frame, the it must have unnamed columns containing information as in the file method; to see an example, type the following in an R console.

system.file("extdata", "dictionary_sbe.csv", package = "oce")

debug

an integer specifying whether debugging information is to be printed during the processing. This is a general parameter that is used by many oce functions. Generally, setting debug=0 turns off the printing, while higher values suggest that more information be printed. If one function calls another, it usually reduces the value of debug first, so that a user can often obtain deeper debugging by specifying higher debug values.

Details

The dictionary format, whether read from a built-in CSV file, or from a user-supplied CSV file, or as a data frame, is as follows.

A character value that specifies the original name of a variable in the data slot of x. This is used in matching such names against targets. Matches may be in the form of equality, or string match (in which any # characters in the column are matched to single digits in variable names).

The oce-convention name to be used for a match.

The unit for the column (only used if the object does not already hold a unit for this entry).

The scale for the column (again, only used if the object does not already hold a scale).

For example, an entry in a CSV file

PSALST##,salinity,,PSS-78

specifies that a variable named "PSALT" followed by 2 digits is to be renamed as "salinity", that the unit (if not already defined within x) is to be blank, and that the scale (again, if not already defined) is to be "PSS-78".

Note that after the renaming is done, duplicate names are handled in the standard oce way, e.g. if the object holds three temperature variables, they will be named temperature, temperature2 and temperature3.

History and Plans

This function was written in late September, 2024. It is likely to evolve through the remaining months of 2024, after real-world testing by the developers.

Author

Dan Kelley

Examples

library(oce)
# Example 1: made-up data
d <- new("oce")
d <- oceSetData(d, "S", c(30, 31))
d <- oceSetData(d, "T", c(10, 11))
dictText <- "S,salinity,,
T,temperature,degree*C,ITS-90"
dictionary <- read.csv(text = dictText, header = FALSE)
rename(d, dictionary)
#> oce object has data as follows.
#>    salinity[1:2]: 30, 31
#>    temperature[1:2]: 10, 11
#
# Example 2: a CIOOS NetCDF file. Note that this file
# is downloaded and removed at the end; in practice,
# it is likely that the file might be retained locally.
if (requireNamespace("curl")) {
    file <- tempfile(fileext = ".nc") # removed later
    server <- "https://cioosatlantic.ca/erddap/files"
    program <- "bio_atlantic_zone_monitoring_program_ctd"
    subprogram <- "Bedford%20Basin%20Monitoring%20Program"
    year <- "2023"
    cast <- "001"
    url <- paste0(
        server, "/", program, "/", subprogram, "/", year,
        "/CTD_BCD2023667_", cast, "_1_DN.ODF.nc"
    )
    curl::curl_download(url, file)
    d <- read.netcdf(file)
    summary(d)
    dd <- rename(d, "ioos")
    summary(dd)
    unlink(file)
}
#> * Time: 2023-01-05 13:14:45
#> * Data Overview
#> 
#>                               Min.                Mean                Max.                Dim. NAs OriginalName      
#>     measurement_time          2023-01-05 13:20:37 2023-01-05 13:22:01 2023-01-05 13:23:26 141  0   "measurement_time"
#>     ScanNumber                5634                6989.8              8352                141  0   "ScanNumber"      
#>     QCNTR_01                  1                   1                   1                   141  0   "QCNTR_01"        
#>     PRESPR01 [dbar]           0.5                 35.5                70.5                141  0   "PRESPR01"        
#>     QPRES_01                  1                   1                   1                   141  0   "QPRES_01"        
#>     TEMPS901 [°C, ITS-90]     5.6177              7.5868              7.8938              141  0   "TEMPS901"        
#>     TEMPP901 [°C, ITS-90]     5.6177              7.5868              7.8938              141  0   "TEMPP901"        
#>     TEMPPR01 [°C, ITS-90]     5.6177              7.5868              7.8938              141  0   "TEMPPR01"        
#>     QTEMP_01                  1                   1                   1                   141  0   "QTEMP_01"        
#>     CNDCST01 [S/m]            2.8712              3.2301              3.2974              141  0   "CNDCST01"        
#>     QCNDC_01                  1                   1                   1                   141  0   "QCNDC_01"        
#>     OXYOCPVL01 [V]            0.816               2.2953              3.07                141  0   "OXYOCPVL01"      
#>     QOXYV_01                  1                   1                   1                   141  0   "QOXYV_01"        
#>     CPHLPR01 [mg/m³]          1.1592              1.6785              4.3245              141  0   "CPHLPR01"        
#>     QCPHLPR01                 1                   1                   1                   141  0   "QCPHLPR01"       
#>     PHXXZZ01                  8.187               8.281               8.784               141  0   "PHXXZZ01"        
#>     QPHPH_01                  1                   1                   1                   141  0   "QPHPH_01"        
#>     IRRDUV01 [μEinstein/s/m²] 0.00025             3.1998              69.881              141  5   "IRRDUV01"        
#>     QPSAR_01                  1                   1.2837              9                   141  0   "QPSAR_01"        
#>     PSALST01 [PSS-78]         29.012              31.173              31.624              141  0   "PSALST01"        
#>     PSLTZZ01 [PSS-78]         29.012              31.173              31.624              141  0   "PSLTZZ01"        
#>     QPSAL_01                  1                   1                   1                   141  0   "QPSAL_01"        
#>     POTMCV01 [°C, ITS-90]     5.6177              7.5835              7.888               141  0   "POTMCV01"        
#>     QPOTM_01                  1                   1                   1                   141  0   "QPOTM_01"        
#>     SIGTEQ01 [kg/m³]          22.852              24.325              24.644              141  0   "SIGTEQ01"        
#>     QSIGP_01                  1                   1                   1                   141  0   "QSIGP_01"        
#>     DOXYZZ01 [ml/l]           0.7969              4.5755              7.178               141  0   "DOXYZZ01"        
#>     QDOXY_01                  1                   1                   1                   141  0   "QDOXY_01"        
#>     RecPerBin                 11                  18.27               110                 141  0   "RecPerBin"       
#>     QCNTR_02                  1                   1                   1                   141  0   "QCNTR_02"        
#>     QCFF_01                   0                   0                   0                   141  0   "QCFF_01"         
#>     time                      2023-01-05 13:14:45 2023-01-05 13:14:45 2023-01-05 13:14:45 1    0   "time"            
#>     latitude [°N]             44.683              44.683              44.683              1    0   "latitude"        
#>     longitude [°E]            -63.633             -63.633             -63.633             1    0   "longitude"       
#> 
#> * Processing Log
#> 
#>     - 2024-09-30 18:29:17 UTC: `Create oce object`
#>     - 2024-09-30 18:29:18 UTC: `read.netcdf("/var/folders/8b/l4h64m1j22v5pb7vj049ff140000gn/T//RtmpSWAQ2t/filefed2367c82a0.nc")`
#> * Time: 2023-01-05 13:14:45
#> * Data Overview
#> 
#>                               Min.                Mean                Max.                Dim. NAs OriginalName      
#>     timeMeasurement           2023-01-05 13:20:37 2023-01-05 13:22:01 2023-01-05 13:23:26 141  0   "measurement_time"
#>     scan                      5634                6989.8              8352                141  0   "ScanNumber"      
#>     pressure [dbar]           0.5                 35.5                70.5                141  0   "PRESPR01"        
#>     temperature [°C, ITS-90]  5.6177              7.5868              7.8938              141  0   "TEMPS901"        
#>     temperature2 [°C, ITS-90] 5.6177              7.5868              7.8938              141  0   "TEMPP901"        
#>     temperature3 [°C, ITS-90] 5.6177              7.5868              7.8938              141  0   "TEMPPR01"        
#>     conductivity [S/m]        2.8712              3.2301              3.2974              141  0   "CNDCST01"        
#>     oxygenVoltage [V]         0.816               2.2953              3.07                141  0   "OXYOCPVL01"      
#>     chlorophyllA [mg/m³]      1.1592              1.6785              4.3245              141  0   "CPHLPR01"        
#>     pH                        8.187               8.281               8.784               141  0   "PHXXZZ01"        
#>     PAR [μEinstein/s/m²]      0.00025             3.1998              69.881              141  5   "IRRDUV01"        
#>     salinity [PSS-78]         29.012              31.173              31.624              141  0   "PSALST01"        
#>     salinity2 [PSS-78]        29.012              31.173              31.624              141  0   "PSLTZZ01"        
#>     theta [°C, ITS-90]        5.6177              7.5835              7.888               141  0   "POTMCV01"        
#>     sigmaTheta [kg/m³]        22.852              24.325              24.644              141  0   "SIGTEQ01"        
#>     oxygen [ml/l]             0.7969              4.5755              7.178               141  0   "DOXYZZ01"        
#>     recordsPerBin             11                  18.27               110                 141  0   "RecPerBin"       
#>     quality                   0                   0                   0                   141  0   "QCFF_01"         
#>     time                      2023-01-05 13:14:45 2023-01-05 13:14:45 2023-01-05 13:14:45 1    0   "time"            
#>     latitude [°N]             44.683              44.683              44.683              1    0   "latitude"        
#>     longitude [°E]            -63.633             -63.633             -63.633             1    0   "longitude"       
#> 
#> * Data-quality Flags
#> 
#>     scan:          "1" 141
#>     pressure:      "1" 141
#>     temperature:   "1" 141
#>     conductivity:  "1" 141
#>     oxygenVoltage: "1" 141
#>     chlorophyllA:  "1" 141
#>     pH:            "1" 141
#>     PAR:           "1" 136, "9" 5
#>     salinity:      "1" 141
#>     theta:         "1" 141
#>     sigmaTheta:    "1" 141
#>     oxygen:        "1" 141
#> 
#> * Processing Log
#> 
#>     - 2024-09-30 18:29:17 UTC: `Create oce object`
#>     - 2024-09-30 18:29:18 UTC: `read.netcdf("/var/folders/8b/l4h64m1j22v5pb7vj049ff140000gn/T//RtmpSWAQ2t/filefed2367c82a0.nc")`
#>     - 2024-09-30 18:29:18 UTC: `rename(..., "/private/var/folders/8b/l4h64m1j22v5pb7vj049ff140000gn/T/RtmpPiqp0d/temp_libpathf07c589698a4/oce/extdata/dictionary_ioos.csv")`