Extract Something From an amsr Object

Extract something from the metadata or data slot of an amsr object.

The [[ method works for all oce objects, i.e. objects inheriting from oce. The purpose is to insulate users from the internal details of oce objects, by looking for items within the various storage slots of the object. Items that are not actually stored in the object can also be extracted, including derived data such as potential temperature, the units of measurement for the data, data-quality flags, etc.

# S4 method for amsr
[[(x, i, j, ...)

Arguments

x
an amsr object.

i
Character string indicating the name of an item to extract.

j
Optional additional information on the i item.

...
Optional additional information (ignored).

Value

In all cases, the returned value is a matrix with with NA values inserted at locations where the raw data equal as.raw(251:255), as explained in “Details”.

Details

Partial matches for i are permitted for metadata, and j is ignored for metadata.

Data within the data slot may be found directly, e.g. j="SSTDay" will yield sea-surface temperature in the daytime satellite, and j="SSTNight" is used to access the nighttime data. In addition, j="SST" yields an average of the night and day values (using just one of these, if the other is missing). This scheme works for all the data stored in amsr objects, namely: time, SST, LFwind, MFwind, vapor, cloud and rain. In each case, the default is to calculate values in scientific units, unless j="raw", in which case the raw data are returned.

The "raw" mode can be useful in decoding the various types of missing value that are used by amsr data, namely as.raw(255) for land, as.raw(254) for a missing observation, as.raw(253) for a bad observation, as.raw(252) for sea ice, or as.raw(251) for missing SST due to rain or missing water vapour due to heavy rain. Note that something special has to be done for e.g. d[["SST","raw"]] because the idea is that this syntax (as opposed to specifying "SSTDay") is a request to try to find good data by looking at both the Day and Night measurements. The scheme employed is quite detailed. Denote by "A" the raw value of the desired field in the daytime pass, and by "B" the corresponding value in the nighttime pass. If either A or B is 255, the code for land, then the result will be 255. If A is 254 (i.e. there is no observation), then B is returned, and the reverse holds also. Similarly, if either A or B equals 253 (bad observation), then the other is returned. The same is done for code 252 (ice) and code 251 (rain).

A two-step process is used to try to find the requested information. First, a class-specific function tries to find it, but if that fails, then a general function is used (see next section).

Details of the general method

If the specialized method produces no matches, the following generalized method is applied. As with the specialized method, the procedure hinges first on the values of i and, optionally, j. The work proceeds in steps, by testing a sequence of possible conditions in sequence.

A check is made as to whether i names one of the standard oce slots. If so, [[ returns the slot contents of that slot. Thus, x[["metadata"]] will retrieve the metadata slot, while x[["data"]] and x[["processingLog"]] return those slots.

If i is a string ending in the "Unit", then the characters preceding that string are taken to be the name of an item in the data object, and a list containing the unit is returned. This list consists of an item named unit, which is an expression(), and an item named scale, which is a string describing the measurement scale. If the string ends in " unit", e.g. x[["temperature unit"]] (note the space), then just the expression is returned, and if it ends in " scale", then just the scale is returned.

If i is a string ending in "Flag", then the corresponding data-quality flag is returned (or NULL if there is no such flag). For example, x[["salinityFlag"]] returns a vector of salinity flags if x is a ctd object.

If i is "sigmaTheta", then the value of swSigmaTheta() is called with x as the sole argument, and the results are returned. Similarly, swSigma0() is used if i="sigma0", and swSpice() is used if i="spice". Of course, these actions only make sense for objects that contain the relevant items within their data slot.

After these possibilities are eliminated, the action depends on whether j has been provided. If j is not provided, or is the string "", then i is sought in the metadata slot, and then in the data slot, returning whichever is found first. In other words, if j is not provided, the metadata slot takes preference over the data slot. However, if j is provided, then it must be either the string "metadata" or "data", and it directs where to look.

If none of the above-listed conditions holds, then NULL is returned, without the issuance of a warning or error message. (This silent operation is employed so that [[ will behave like the normal R version.)

See also

Examples

if (FALSE) {
# Show a daytime SST image, along with an indication of whether
# the NA values are from rain.
library(oce)
earth <- read.amsr("f34_20160102v7.2.gz")
fclat <- subset(earth , 35 <= latitude & latitude <= 55)
fc <- subset(fclat , -70 <= longitude & longitude <= -30)
par(mfrow=c(2, 1))
plot(fc, "SSTDay")
rainy <- fc[["SSTDay", "raw"]] == as.raw(0xfb)
lon <- fc[["longitude"]]
lat <- fc[["latitude"]]
asp <- 1 / cos(pi*mean(lat)/180)
imagep(lon, lat, rainy, asp=asp)
mtext("red: too rainy to sense SSTDay")
}