Summarize an oce Object

Provide a textual summary of some pertinent aspects of the object, including selected components of its metadata slot, statistical and dimensional information on the entries in the data slot, and a listing of the contents of its processingLog slot. The details depend on the class of the object, especially for the metadata slot, so it can help to consult the specialized documentation, e.g. summary,ctd-method for CTD objects (i.e. objects inheriting from the ctd class.) It is important to note that this is not a good way to learn the details of the object contents. Instead, for an object named object, say, one might use str(object) to learn about all the contents, or str(object[["metadata"]]) to learn about the metadata, etc.

# S4 method for oce
summary(object, ...)

Arguments

object

The object to be summarized.

...

Extra arguments (ignored)

Examples

o <- new("oce")
summary(o)
#> * Processing Log
#> 
#>     - 2022-07-04 09:25:30 UTC: `Create oce object`