Summarize a sealevel Object

Summarizes some of the data in a sealevel object.

Usage

# S4 method for class 'sealevel'
summary(object, ...)

Arguments

object

A sealevel object.

...

further arguments passed to or from other methods.

Value

A matrix containing statistics of the elements of the data slot.

See also

Author

Dan Kelley

Examples

library(oce)
data(sealevel)
summary(sealevel)
#> Sealevel Summary
#> ----------------
#> 
#> * number:              490
#> * name:                HALIFAX
#> * sampling delta-t:    1 hour
#> * Location:            44.667N 63.583W 
#> * year:                2003
#> * number of observations:   6659 
#> *    "      non-missing:    6659 
#> * Time: 2003-01-01 13:00:00 to 2003-10-08 11:00:00 (6659 samples, mean increment 1.009012 hour)
#> * Data Overview
#> 
#>                   Min.                Mean                Max.                Dim. NAs
#>     elevation [m] 0                   0.98622             2.84                6659 0  
#>     time          2003-01-01 13:00:00 2003-05-21 08:21:11 2003-10-08 11:00:00 6659 0  
#> 
#> * Processing Log
#> 
#>     - 2016-05-05 18:33:53 UTC: `create 'sealevel' object`
#>     - 2016-05-05 18:33:53 UTC: `read.sealevel(file="490-01-JAN-2003_slev.csv", tz="UTC")`
#>     - 2016-05-05 18:33:53 UTC: `oce.edit(x = sealevel, item = "longitude", value = -sealevel[["longitude"]],     reason = "Fix longitude hemisphere")`