Seawater absolute salinity, in GSW formulation

Compute the seawater Absolute Salinity, according to the GSW/TEOS-10 formulation with gsw::gsw_SA_from_SP() in the gsw package. Typically, this is a fraction of a unit higher than practical salinity as defined in the UNESCO formulae.

swAbsoluteSalinity(
  salinity,
  pressure = NULL,
  longitude = NULL,
  latitude = NULL
)

Arguments

salinity
either practical salinity (in which case temperature and pressure must be provided) or an oce object (in which case salinity, etc. are inferred from the object).

pressure
pressure in dbar.

longitude
longitude of observation.

latitude
latitude of observation.

Value

Absolute Salinity in \(g/kg\).

References

McDougall, T.J. and P.M. Barker, 2011: Getting started with TEOS-10 and the Gibbs Seawater (GSW) Oceanographic Toolbox, 28pp., SCOR/IAPSO WG127, ISBN 978-0-646-55621-5.

See also

Examples

if (FALSE) {
sa <- swAbsoluteSalinity(35.5, 300, 260, 16)
stopifnot(abs(35.671358392019094 - sa) < 00.000000000000010)
}