Seawater haline contraction coefficient

Compute \(\beta\), the haline contraction coefficient for seawater.

swBeta(
  salinity,
  temperature = NULL,
  pressure = 0,
  longitude = NULL,
  latitude = NULL,
  eos = getOption("oceEOS", default = "gsw")
)

Arguments

salinity
either practical salinity (in which case temperature and pressure must be provided) or an oce object (in which case salinity, etc. are inferred from the object).

temperature
in-situ temperature (\(^\circ\)C), defined on the ITS-90 scale; see “Temperature units” in the documentation for swRho().

pressure
seawater pressure (dbar)

longitude
longitude of observation (only used if eos="gsw"; see ‘Details’).

latitude
latitude of observation (only used if eos="gsw"; see ‘Details’).

eos
equation of state, either "unesco" or "gsw".

Value

Value in 1/psu.

References

The eos="unesco" formulae are based on the UNESCO equation of state, but are formulated empirically by Trevor J. McDougall, 1987, Neutral Surfaces, Journal of Physical Oceanography, volume 17, pages 1950-1964. The eos="gsw" formulae come from GSW; see references in the swRho() documentation.

See also

Author

Dan Kelley