Electrical Conductivity Ratio From Salinity, Temperature and Pressure

Compute electrical conductivity ratio based on salinity, temperature, and pressure (relative to the conductivity of seawater with salinity=35, temperature68=15, and pressure=0).

Usage

swCSTp(
  salinity,
  temperature = 15,
  pressure = 0,
  eos = getOption("oceEOS", default = "gsw")
)

Arguments

salinity

practical salinity, or a CTD object (in which case its temperature and pressure are used, and the next two arguments are ignored)

temperature

in-situ temperature (\(^\circ\)C), defined on the ITS-90 scale; see the examples, as well as the “Temperature units” section in the documentation for swRho().

pressure

pressure (dbar)

eos

equation of state, either "unesco" or "gsw".

Value

Conductivity ratio (unitless), i.e. the ratio of conductivity to the conductivity at salinity=35, temperature=15 (IPTS-68 scale) and pressure=0, which has numerical value 42.9140 mS/cm = 4.29140 S/m (see Culkin and Smith, 1980, in the regression result cited at the bottom of the left-hand column on page 23).

Details

If eos="unesco", the calculation is done by a bisection root search on the UNESCO formula relating salinity to conductivity, temperature, and pressure (see swSCTp()). If it is "gsw" then the Gibbs-SeaWater formulation is used, via gsw::gsw_C_from_SP().

References

Fofonoff, P. and R. C. Millard Jr, 1983. Algorithms for computation of fundamental properties of seawater. Unesco Technical Papers in Marine Science, 44, 53 pp.

Culkin, F., and Norman D. Smith, 1980. Determination of the concentration of potassium chloride solution having the same electrical conductivity, at 15 C and infinite frequency, as standard seawater of salinity 35.0000 ppt (Chlorinity 19.37394 ppt). IEEE Journal of Oceanic Engineering, 5, pp 22-23.

See also

Author

Dan Kelley

Examples

stopifnot(abs(1.0 - swCSTp(35, T90fromT68(15), 0, eos = "unesco")) < 1e-7)
stopifnot(abs(1.0 - swCSTp(34.25045, T90fromT68(15), 2000, eos = "unesco")) < 1e-7)
stopifnot(abs(1.0 - swCSTp(34.25045, T90fromT68(15), 2000, eos = "gsw")) < 1e-7)