Water Depth

Retrieve or compute depth below the surface, i.e. a positive number within the water column. If the first parameter is an oce object that has an element named "depth" in its data slot, then return that value. Otherwise, compute depth from a formula that includes pressure and latitude, as explained in “Details”.

Usage

swDepth(
  pressure,
  latitude = 45,
  eos = getOption("oceEOS", default = "gsw"),
  debug = getOption("oceDebug")
)

Arguments

pressure

either pressure (dbar), in which case latitude must also be given, or a ctd object, in which case latitude will be inferred from the object.

latitude

numeric value for latitude in degrees North.

eos

character value indicating the formulation to be used, either "unesco" or "gsw".

debug

an integer specifying whether debugging information is to be printed during the processing. This is a general parameter that is used by many oce functions. Generally, setting debug=0 turns off the printing, while higher values suggest that more information be printed. If one function calls another, it usually reduces the value of debug first, so that a user can often obtain deeper debugging by specifying higher debug values.

Value

swDepth returns depth below the ocean surface, in metres.

Details

For the calculated case, the method depends on the value of eos parameter. If this is "unesco", then depth is calculated from pressure using Saunders and Fofonoff's method, with the formula refitted for 1980 UNESCO equation of state (reference 1). On the other hand, if it is eos="gsw", then gsw::gsw_z_from_p() from the gsw package (references 2 and 3) is used.

References

Unesco 1983. Algorithms for computation of fundamental properties of seawater, 1983. Unesco Tech. Pap. in Mar. Sci., No. 44, 53 pp.

IOC, SCOR, and IAPSO (2010). The international thermodynamic equation of seawater-2010: Calculation and use of thermodynamic properties. Technical Report 56, Intergovernmental Oceanographic Commission, Manuals and Guide.

McDougall, T.J. and P.M. Barker, 2011: Getting started with TEOS-10 and the Gibbs Seawater (GSW) Oceanographic Toolbox, 28pp., SCOR/IAPSO WG127, ISBN 978-0-646-55621-5.

See also

Author

Dan Kelley

Examples

d <- swDepth(10, 45)