Seawater lapse rate

Compute adiabatic lapse rate

swLapseRate(
  salinity,
  temperature = NULL,
  pressure = NULL,
  longitude = NULL,
  latitude = NULL,
  eos = getOption("oceEOS", default = "gsw")
)

Arguments

salinity
either salinity (PSU) (in which case temperature and pressure must be provided) or a ctd object (in which case salinity, temperature and pressure are determined from the object, and must not be provided in the argument list).

temperature
in-situ temperature (\(^\circ\)C), defined on the ITS-90 scale; see “Temperature units” in the documentation for swRho().

pressure
pressure (dbar)

longitude
longitude of observation (only used if eos="gsw"; see ‘Details’).

latitude
latitude of observation (only used if eos="gsw"; see ‘Details’).

eos
equation of state, either "unesco" (references 1 and 2) or "gsw" (references 3 and 4).

Value

Lapse rate (\(deg\)C/m).

Details

If eos="unesco", the density is calculated using the UNESCO equation of state for seawater (references 1 and 2), and if eos="gsw", the GSW formulation (references 3 and 4) is used.

References

Fofonoff, P. and R. C. Millard Jr, 1983. Algorithms for computation of fundamental properties of seawater. Unesco Technical Papers in Marine Science, 44, 53 pp. (Section 7, pages 38-40)

See also

Examples

lr <- swLapseRate(40, 40, 10000) # 3.255976e-4