Seawater Salinity From Temperature and Density

Compute Practical or Absolute Salinity, given in-situ or Conservative Temperature, density, and pressure. This is mainly used to draw isopycnal lines on TS diagrams, hence the dual meanings for salinity and temperature, depending on the value of eos.

Usage

swSTrho(
  temperature,
  density,
  pressure,
  eos = getOption("oceEOS", default = "gsw")
)

Arguments

temperature

in-situ temperature (\(^\circ\)C), defined on the ITS-90 scale; see “Temperature units” in the documentation for swRho().

density

in-situ density or sigma value (\(kg/m^3\))

pressure

in-situ pressure (dbar)

eos

equation of state, either "unesco" (see references 1 and 2) or "gsw" (see references 3 and 4).

Value

Practical Salinity, if eos="unesco", or Absolute Salinity, if eos="gsw".

Details

For eos="unesco", finds the practical salinity that yields the given density, with the given in-situ temperature and pressure. The method is a bisection search with a salinity tolerance of 0.001. For eos="gsw", the function gsw::gsw_SA_from_rho() in the gsw package is used to infer Absolute Salinity from Conservative Temperature.

References

Fofonoff, P. and R. C. Millard Jr, 1983. Algorithms for computation of fundamental properties of seawater. Unesco Technical Papers in Marine Science, 44, 53 pp

Gill, A.E., 1982. Atmosphere-ocean Dynamics, Academic Press, New York, 662 pp.

IOC, SCOR, and IAPSO (2010). The international thermodynamic equation of seawater-2010: Calculation and use of thermodynamic properties. Technical Report 56, Intergovernmental Oceanographic Commission, Manuals and Guide.

McDougall, T.J. and P.M. Barker, 2011: Getting started with TEOS-10 and the Gibbs Seawater (GSW) Oceanographic Toolbox, 28pp., SCOR/IAPSO WG127, ISBN 978-0-646-55621-5.

See also

Author

Dan Kelley

Examples

swSTrho(10, 22, 0, eos = "gsw") # 28.76285
#> [1] 28.76249
swSTrho(10, 22, 0, eos = "unesco") # 28.651625
#> [1] 28.65163