Seawater density anomaly

Compute \(\sigma_\theta\), the density of seawater, minus 1000 kg/m\(^3\).

swSigma(
  salinity,
  temperature = NULL,
  pressure = NULL,
  longitude = NULL,
  latitude = NULL,
  eos = getOption("oceEOS", default = "gsw")
)

Arguments

salinity

either practical salinity (in which case temperature and pressure must be provided) or an oce object, in which case salinity, temperature (in the ITS-90 scale; see next item), etc. are inferred from the object, ignoring the other parameters, if they are supplied.

temperature

in-situ temperature (\(^\circ\)C), defined on the ITS-90 scale. This scale is used by GSW-style calculation (as requested by setting eos="gsw"), and is the value contained within ctd objects (and probably most other objects created with data acquired in the past decade or two). Since the UNESCO-style calculation is based on IPTS-68, the temperature is converted within the present function, using T68fromT90().

pressure

pressure (dbar)

longitude

longitude of observation (only used if eos="gsw"; see ‘Details’).

latitude

latitude of observation (only used if eos="gsw"; see ‘Details’).

eos

equation of state, either "unesco" (references 1 and 2) or "gsw" (references 3 and 4).

Value

Density anomaly (kg/m\(^3\)), as computed with swRho(), minus- 1000 kg/m\(^3\).

References

See citations provided in the swRho() documentation.

See also

Author

Dan Kelley

Examples

library(oce)
swSigma(35, 13, 1000, longitude=300, latitude=30, eos="gsw") # 30.82374
#> [1] 30.82374
swSigma(35, T90fromT68(13), 1000, eos="unesco")    # 30.8183
#> [1] 30.8183