Seawater Temperature from Salinity and Density

Compute in-situ temperature, given salinity, density, and pressure.

Usage

swTSrho(
  salinity,
  density,
  pressure = NULL,
  eos = getOption("oceEOS", default = "gsw")
)

Arguments

salinity

in-situ salinity (PSU)

density

in-situ density or sigma value (kg/m\(^3\))

pressure

in-situ pressure (dbar)

eos

equation of state to be used, either "unesco" or "gsw" (ignored at present).

Value

In-situ temperature in \(^\circ\)C on the ITS-90 scale.

Details

Finds the temperature that yields the given density, with the given salinity and pressure. The method is a bisection search with temperature tolerance 0.001 \(^\circ\)C.

References

Fofonoff, P. and R. C. Millard Jr, 1983. Algorithms for computation of fundamental properties of seawater. Unesco Technical Papers in Marine Science, 44, 53 pp

Gill, A.E., 1982. Atmosphere-ocean Dynamics, Academic Press, New York, 662 pp.

See also

Author

Dan Kelley

Examples

swTSrho(35, 23, 0, eos = "unesco") # 26.11301
#> [1] 26.10677