Seawater Thermal Conductivity

Compute seawater thermal conductivity, in \(W m^{-1\circ}C^{-1}\)

Usage

swThermalConductivity(salinity, temperature = NULL, pressure = NULL)

Arguments

salinity

salinity (PSU), or a ctd object, in which case temperature and pressure will be ignored.

temperature

in-situ temperature (\(^\circ\)C), defined on the ITS-90 scale; see “Temperature units” in the documentation for swRho().

pressure

pressure (dbar)

Value

Conductivity of seawater in \(W m^{-1} {^\circ} C^{-1}\). To calculate thermal diffusivity in \(m^2/s^2\), divide by the product of density and specific heat, as in the example.

Details

Caldwell's (1974) detailed formulation is used. To be specific, his equation 6 to calculate K, and his two sentences above that equation are used to infer this to be K(0,T,S) in his notation of equation 7. Then, application of his equations 7 and 8 is straightforward. He states an accuracy for this method of 0.3 percent. (See the check against his Table 1 in the “Examples”.)

References

Caldwell, Douglas R., 1974. Thermal conductivity of seawater, Deep-sea Research, 21, 131-137.

See also

Author

Dan Kelley

Examples

library(oce)
# Values in m^2/s, a unit that is often used instead of W/(m*degC).
swThermalConductivity(35, 10, 100) / (swRho(35, 10, 100) * swSpecificHeat(35, 10, 100)) # ocean
#> [1] 1.512538e-07
swThermalConductivity(0, 20, 0) / (swRho(0, 20, 0) * swSpecificHeat(0, 20, 0)) # lab
#> [1] 1.526025e-07
# Caldwell Table 1 gives 1478e-6 cal/(cm*sec*degC) at 31.5 o/oo, 10degC, 1kbar
joulePerCalorie <- 4.18400
cmPerM <- 100
swThermalConductivity(31.5, 10, 1000) / joulePerCalorie / cmPerM
#> [1] 0.00147754