Seawater Viscosity

Compute viscosity of seawater, in \(Pa\cdot s\)

Usage

swViscosity(salinity, temperature)

Arguments

salinity

either salinity (PSU) (in which case temperature and pressure must be provided) or a ctd object (in which case salinity, temperature and pressure are determined from the object, and must not be provided in the argument list).

temperature

in-situ temperature (\(^\circ\)C), defined on the ITS-90 scale; see “Temperature units” in the documentation for swRho(), and the examples below.

Value

Viscosity of seawater in \(Pa\cdot s\). Divide by density to get kinematic viscosity in \(m^2/s\).

Details

If the first argument is a ctd object, then salinity, temperature and pressure values are extracted from it, and used for the calculation.

The result is determined from a regression of the data provided in Table 87 of Dorsey (1940). The fit matches the table to within 0.2 percent at worst, and with average absolute error of 0.07 percent. The maximum deviation from the table is one unit in the last decimal place.

No pressure dependence was reported by Dorsey (1940).

References

N. Ernest Dorsey (1940), Properties of ordinary Water-substance, American Chemical Society Monograph Series. Reinhold Publishing.

See also

Author

Dan Kelley

Examples

swViscosity(30, 10) # 0.001383779
#> [1] 0.001383779