Vertical coordinate

Compute height above the surface. This is the negative of depth, and so is defined simply in terms of swDepth().

swZ(pressure, latitude = 45, eos = getOption("oceEOS", default = "gsw"))

Arguments

pressure

either pressure (dbar), in which case lat must also be given, or a ctd object, in which case lat will be inferred from the object.

latitude

Latitude in \(^\circ\)N or radians north of the equator.

eos

indication of formulation to be used, either "unesco" or "gsw".

See also