Astronomical Calculations for Tidem

Do some astronomical calculations for tidem(). This function is based directly on t_astron in the T_TIDE Matlab package (see Pawlowicz et al. 2002), which inherits from the Fortran code described by Foreman (1978).

tidemAstron(t)

Arguments

t
Either a time in POSIXct format (with "UTC" timezone, or else odd behaviours may result), or an integer. In the second case, it is converted to a time with numberAsPOSIXct(), using tz="UTC".

Value

A list containing items named astro and ader (see the T_TIDE documentation).

References

Foreman, M. G. G., 1978. Manual for Tidal Currents Analysis and Prediction. Pacific Marine Science Report. British Columbia, Canada: Institute of Ocean Sciences, Patricia Bay.

Pawlowicz, Rich, Bob Beardsley, and Steve Lentz, 2002. Classical tidal harmonic analysis including error estimates in MATLAB using T_TIDE. Computers and Geosciences, 28, 929-937.

See also

Examples

tidemAstron(as.POSIXct("2008-01-22 18:50:24"))

#> $astro
#> [1] 1.44963597 0.34000639 0.83797569 0.14240012 0.08562115 0.78633081
#> 
#> $ader
#> [1] 9.661368e-01 3.660110e-02 2.737909e-03 3.094541e-04 1.470939e-04
#> [6] 1.308208e-07
#>