Convert an adp Object to ENU Coordinates

Convert an adp Object to ENU Coordinates

Usage

toEnuAdp(x, declination = 0, debug = getOption("oceDebug"))

Arguments

x

an adp object.

declination

magnetic declination to be added to the heading, to get ENU with N as "true" north.

debug

an integer specifying whether debugging information is to be printed during the processing. This is a general parameter that is used by many oce functions. Generally, setting debug=0 turns off the printing, while higher values suggest that more information be printed. If one function calls another, it usually reduces the value of debug first, so that a user can often obtain deeper debugging by specifying higher debug values.

References

https://nortek.zendesk.com/hc/en-us/articles/360029820971-How-is-a-Coordinate-transformation-done-

See also

Author

Dan Kelley