Interpolate Within a topo Object

Bilinear interpolation is used so that values will vary smoothly within a longitude-latitude grid cell. Note that the sign convention for longitude and latitude must match that in topo.

Usage

topoInterpolate(longitude, latitude, topo)

Arguments

longitude

Vector of longitudes (in the same sign convention as used in topo).

latitude

Vector of latitudes (in the same sign convention as used in topo).

topo

A topo object.

Value

Vector of heights giving the elevation of the earth above means sea level at the indicated location on the earth.

See also

Author

Dan Kelley

Examples

library(oce)
data(topoWorld)
# "The Gully", approx. 400m deep, connects Gulf of St Lawrence with North Atlantic
topoInterpolate(45, -57, topoWorld)
#> [1] -5208.5