Convert UTM to Longitude and Latitude

Convert UTM to Longitude and Latitude

utm2lonlat(easting, northing, zone = 1, hemisphere = "N", km = FALSE)

Arguments

easting

easting coordinate (in km or m, depending on value of km). Alternatively, a list containing items named easting, northing, and zone, in which case these are taken from the list and the arguments named northing, zone and are ignored.

northing

northing coordinate (in km or m, depending on value of km).

zone

UTM zone

hemisphere

indication of hemisphere; "N" for North, anything else for South.

km

logical value indicating whether easting and northing are in kilometers or meters.

Value

A list containing longitude and latitude.

References

https://en.wikipedia.org/wiki/Universal_Transverse_Mercator_coordinate_system, downloaded May 31, 2014.

See also

Author

Dan Kelley

Examples

library(oce)
## Cape Split, in the Minas Basin of the Bay of Fundy
utm2lonlat(852863, 5029997, 19)
#> $longitude
#> [1] -64.49657
#> 
#> $latitude
#> [1] 45.33462
#>