Write a CTD Data Object as a CSV File

Writes a comma-separated file containing the data frame stored in the data slot of the first argument. The file is suitable for reading with a spreadsheet, or with read.csv(). This output file will contain some of the metadata in x, if metadata is TRUE.

write.ctd(object, file, metadata = TRUE, flags = TRUE, format = "csv")

Arguments

object

a ctd object.

file

Either a character string (the file name) or a connection. If not provided, file defaults to stdout().

metadata

a logical value indicating whether to put some selected metadata elements at the start of the output file.

flags

a logical value indicating whether to show data-quality flags as well as data.

format

string indicating the format to use. This may be "csv" for a simple CSV format, or "whp" for the World Hydrographic Program format, described in reference 1 and exemplified in reference 2.

References

The following links used to work, but failed as of December 2020.

https://www.nodc.noaa.gov/woce/woce_v3/wocedata_1/whp/exchange/exchange_format_desc.htm

https://www.nodc.noaa.gov/woce/woce_v3/wocedata_1/whp/exchange/example_ct1.csv

See also

Author

Dan Kelley

Examples

if (FALSE) {
library(oce)
data(ctd)
write.ctd(ctd, "ctd.csv")
d <- read.csv("ctd.csv")
plot(as.ctd(d$salinity, d$temperature, d$pressure))
}