Class to Store XBT (Expendable Bathythermograph) Data

This class stores expendable bathythermograph (XBT) data, e.g. as read using read.xbt() and related functions. Reference 1 gives some information on Sippican devices, and reference 2 is a useful introduction to the modern literature on XBTs in general.

Slots

data

As with all oce objects, the data slot for xbt objects is a list containing the main data for the object. The key items stored in this slot are depth (or z) and temperature, although some datasets also have soundSpeed. Note that depth and z are inferred from time in water, using an empirical formula for instrument descent rate, and that soundSpeed is calculated using a fixed practical salinity of 35. Note that the [[ accessor will compute any of depth, z or pressure, based on whatever is in the data object. Similarly, soundspeed will compute sound speed (assuming a practical salinity of 35), if that that item is present in the data slot.

metadata

As with all oce objects, the metadata slot for xbt objects is a list containing information about the data or about the object itself.

processingLog

As with all oce objects, the processingLog slot for xbt objects is a list with entries describing the creation and evolution of the object. The contents are updated by various oce functions to keep a record of processing steps. Object summaries and processingLogShow() both display the log.

Modifying slot contents

Although the [[<- operator may permit modification of the contents of xbt objects (see [[<-,xbt-method), it is better to use oceSetData() and oceSetMetadata(), because those functions save an entry in the processingLog that describes the change.

Retrieving slot contents

The full contents of the data and metadata slots of a xbt object may be retrieved in the standard R way using slot(). For example slot(o,"data") returns the data slot of an object named o, and similarly slot(o,"metadata") returns the metadata slot.

The slots may also be obtained with the [[,xbt-method operator, as e.g. o[["data"]] and o[["metadata"]], respectively.

The [[,xbt-method operator can also be used to retrieve items from within the data and metadata slots. For example, o[["temperature"]] can be used to retrieve temperature from an object containing that quantity. The rule is that a named quantity is sought first within the object's metadata slot, with the data slot being checked only if metadata does not contain the item. This [[ method can also be used to get certain derived quantities, if the object contains sufficient information to calculate them. For example, an object that holds (practical) salinity, temperature and pressure, along with longitude and latitude, has sufficient information to compute Absolute Salinity, and so o[["SA"]] will yield the calculated Absolute Salinity.

It is also possible to find items more directly, using oceGetData() and oceGetMetadata(), but neither of these functions can retrieve derived items.

References

Sippican, Inc. "Bathythermograph Data Acquisition System: Installation, Operation and Maintenance Manual (P/N 308195, Rev. A)," 2003. https://pages.uoregon.edu/drt/MGL0910_Science_Report/attachments/MK21_ISA_Manual_Rev_A.pdf.

Cheng, Lijing, John Abraham, Gustavo Goni, Timothy Boyer, Susan Wijffels, Rebecca Cowley, Viktor Gouretski, et al. "XBT Science: Assessment of Instrumental Biases and Errors." Bulletin of the American Meteorological Society 97, no. 6 (June 2016): 924-33. 10.1175/BAMS-D-15-00031.1

See also

Author

Dan Kelley