Decode an item from a Nortek AD2CP file header

Decode an item from a Nortek AD2CP file header

ad2cpHeaderValue(x, key, item, numeric = TRUE, default)

Arguments

x

an adp object that holds AD2CP data.

key

Character value that identifies a particular line in x[["text"]].

item

Character value indicating the name of the item sought.

numeric

Logical value indicating whether to convert the return value from a string to a numerical value.

default

Optional value to be used if the item is not found in the header, or if the header is NULL (as in the case of a split-up file that lacks the initial header information)

Value

String or number interpreted from the x[["text"]], or NULL, if the desired item is not found there, or if x is not of the required class and variety.

See also

Examples

if (FALSE) {
d <- read.oce("a.ad2cp")
# The examples start with the line in x[["text"]][[1]]; note that in the second
# example, it would be insuficient to use a key of "BEAMCFGLIST", because that will
# yield 4 lines, and the function is not designed to handle that.

# ID,STR=\"Signature1000\",SN=123456
type <- ad2cpHeaderValue(d, "ID", "STR", numeric=FALSE)
serialNumber <- ad2cpHeaderValue(d, "ID", "SN")

# BEAMCFGLIST,BEAM=1,THETA=25.00,PHI=0.00,FREQ=1000,BW=25,BRD=1,HWBEAM=1,ZNOM=60.00
beam1Angle <- ad2cpHeaderValue(d, "BEAMCFGLIST,BEAM=1", "THETA")
frequency <- ad2cpHeaderValue(d, "BEAMCFGLIST,BEAM=1", "FREQ", default=NA)
}