Flag adp Data Past Water Column Boundary

Flag variables with the same dimension of v in an adp object that are beyond the water column boundary. Currently, this operation can only be performed on adp objects that contain bottom ranges. Commonly, handleFlags() would then be used to remove such data.

adpFlagPastBoundary(
  x = NULL,
  fields = NULL,
  df = 20,
  trim = 0.15,
  good = 1,
  bad = 4,
  debug = getOption("oceDebug")
)

Arguments

x

an adp object containing bottom ranges.

fields

a variable contained within x indicating which field to flag.

df

the degrees of freedom to use during the smoothing spline operation.

trim

a scale factor for boundary trimming (see “Details”).

good

number stored in flags to indicate good data.

bad

number stored in flags to indicate bad data.

debug

an integer specifying whether debugging information is to be printed during the processing. This is a general parameter that is used by many oce functions. Generally, setting debug=0 turns off the printing, while higher values suggest that more information be printed. If one function calls another, it usually reduces the value of debug first, so that a user can often obtain deeper debugging by specifying higher debug values.

Value

adpFlagPastBoundary returns an adp object with flags adjusted in the specified fields if data are beyond the water column boundary.

Details

This works by fitting a smoothing spline to a bottom range with a defined number of degrees of freedom. For each time, it then searches to determine which associated distances are greater than the predicted smooth spline multiplied by \(1-trim\).

See also

Author

Jaimie Harbin, Clark Richards, and Dan Kelley