Convert Argo Data Name to Oce Name

This function is used internally by read.argo() to convert Argo-convention data names to oce-convention names. Users should not call this directly, since its return value may be changed at any moment (e.g. to include units as well as names).

argoNames2oceNames(names, ignore.case = TRUE)

Arguments

names

vector of character strings containing names in the Argo convention.

ignore.case

a logical value passed to gsub(), indicating whether to ignore the case of input strings. The default is set to TRUE because some data files use lower-case names, despite the fact that the Argo documentation specifies upper-case.

Value

A character vector of the same length as names, but with replacements having been made for all known quantities.

Details

The inference of names was done by inspection of some data files, based on reference 1. It should be noted, however, that the data files examined contain some names that are not documented in reference 1, and others that are listed only in its changelog, with no actual definitions being given. For example, the files had six distinct variable names that seem to relate to phase in the oxygen sensor, but these are not translated by the present function because these variable names are not defined in reference 1, or not defined uniquely in reference 2.

The names are converted with gsub(), using the ignore.case argument of the present function. The procedure is to first handle the items listed in the following table, with string searches anchored to the start of the string. After that, the qualifiers _ADJUSTED, _ERROR and _QC, are translated to Adjusted, Error, and QC, respectively.

Argo name	oce name

BBP	bbp

BETA_BACKSCATTERING	betaBackscattering

BPHASE_OXY	bphaseOxygen

CDOM	CDOM

CNDC	conductivity

CHLA	chlorophyllA

CP	beamAttenuation

CYCLE_NUMBER	cycleNumber (both this and cycle are handled by the [[ operator)

DATA_CENTRE	dataCentre

DATA_MODE	dataMode

DATA_STATE_INDICATOR	dataStateIndicator

DC_REFERENCE	DCReference

DIRECTION	direction

DOWN_IRRADIANCE	downwellingIrradiance

DOWNWELLING_PAR	downwellingPAR

FIRMWARE_VERSION	firmwareVersion

FIT_ERROR_NITRATE	fitErrorNitrate

FLUORESCENCE_CDOM	fluorescenceCDOM

FLUORESCENCE_CHLA	fluorescenceChlorophyllA

INST_REFERENCE	instReference

JULD	juld (and used to compute time)

JULD_QC_LOCATION	juldQCLocation

LATITUDE	latitude

LONGITUDE	longitude

MOLAR_DOXY	oxygenUncompensated

PH_IN_SITU_FREE	pHFree

PH_IN_SITU_TOTAL	pH

PI_NAME	PIName

PLATFORM_NUMBER	id

POSITION_ACCURACY	positionAccuracy

POSITIONING_SYSTEM	positioningSystem

PROFILE	profile

PROJECT_NAME	projectName

RAW_DOWNWELLING_IRRADIANCE	rawDownwellingIrradiance

RAW_DOWNWELLING_PAR	rawDownwellingPAR

RAW_UPWELLING_RADIANCE	rawUpwellingRadiance

STATION_PARAMETERS	stationParameters

TEMP	temperature

TEMP_CPU_CHLA	temperatureCPUChlorophyllA

TEMP_DOXY	temperatureOxygen

TEMP_NITRATE	temperatureNitrate

TEMP_PH	temperaturePH

TEMP_SPECTROPHOTOMETER_NITRATE	temperatureSpectrophotometerNitrate

TILT	tilt

TURBIDITY	turbidity

UP_RADIANCE	upwellingRadiance

UV_INTENSITY	UVIntensity

UV_INTENSITY_DARK_NITRATE	UVIntensityDarkNitrate

UV_INTENSITY_NITRATE	UVIntensityNitrate

VRS_PH	VRSpH

WMO_INST_TYPE	WMOInstType

References

Argo User's Manual Version 3.3, Nov 89th, 2019, available at https://archimer.ifremer.fr/doc/00187/29825/ online.

Argo list of parameters in an excel spreadsheet, available at http://www.argodatamgt.org/content/download/27444/187206/file/argo-parameters-list-core-and-b.xlsx

See also