Coerce data into a CM object

Coerce data into a CM object

as.cm(
  time,
  u = NULL,
  v = NULL,
  pressure = NULL,
  conductivity = NULL,
  temperature = NULL,
  salinity = NULL,
  longitude = NA,
  latitude = NA,
  filename = "",
  debug = getOption("oceDebug")
)

Arguments

time
A vector of times of observation, or an oce object that holds time, in addition to either both u and v, or both directionTrue and speedHorizontal.

u
either a numerical vector containing the eastward component of velocity, in m/s, or an oce object that can can be coerced into a cm object. In the second case, the other arguments to the present function are ignored.

v
vector containing the northward component of velocity in m/s.

pressure
vector containing pressure in dbar. Ignored if the first argument contains an oce object holding pressure.

conductivity
Optional vector of conductivity. Ignored if the first argument contains an oce object holding pressure.

temperature
Optional vector of temperature. Ignored if the first argument contains an oce object holding temperature

salinity
Optional vector of salinity, assumed to be Practical Salinity. Ignored if the first argument contains an oce object holding salinity

longitude
Optional longitude in degrees East. Ignored if the first argument contains an oce object holding longitude.

latitude
Latitude in degrees North. Ignored if the first argument contains an oce object holding latitude.

filename
Optional source file name

debug
an integer specifying whether debugging information is to be printed during the processing. This is a general parameter that is used by many oce functions. Generally, setting debug=0 turns off the printing, while higher values suggest that more information be printed. If one function calls another, it usually reduces the value of debug first, so that a user can often obtain deeper debugging by specifying higher debug values.

See also