World Coastline

This is a coarse resolution coastline at scale 1:110M, with 10,696 points, suitable for world-scale plots plotted at a small size, e.g. inset diagrams. Finer resolution coastline files are provided in the ocedata package.

Source

Downloaded from https://www.naturalearthdata.com, in ne_110m_admin_0_countries.shp in July 2015, with an update on December 16, 2017.

Installing your own datasets

Follow the procedure along the lines described in “Details”, where of course your source file will differ. Also, you should change the name of the coastline object from coastlineWorld, to avoid conflicts with the built-in dataset. Save the .rda file to some directory of your choosing, e.g. perhaps /data/coastlines or ~/data/coastlines on a unix-type machine. Then, whenever you need the file, use load() to load it. Most users find it convenient to do the loading in an Rprofile() startup file.

See also