Add Isopycnal Curves to TS Plot

Adds isopycnal lines to an existing temperature-salinity plot. This is called by plotTS(), and may be called by the user also, e.g. if an image plot is used to show TS data density.

drawIsopycnals(
  nlevels = 6,
  levels,
  rotate = TRUE,
  rho1000 = FALSE,
  digits = 2,
  eos = getOption("oceEOS", default = "gsw"),
  cex = 0.75 * par("cex"),
  col = "darkgray",
  lwd = par("lwd"),
  lty = par("lty")
)

Arguments

nlevels
suggested number of density levels (i.e. isopycnal curves); ignored if levels is supplied.

levels
optional density levels to draw.

rotate
boolean, set to TRUE to write all density labels horizontally.

rho1000
boolean, set to TRUE to write isopycnal labels as e.g. 1024 instead of 24.

digits
number of decimal digits to use in label (supplied to round()).

eos
equation of state to be used, either "unesco" or "gsw".

cex
size for labels.

col
color for lines and labels.

lwd
line width for isopycnal curves

lty
line type for isopycnal curves

Value

None.

See also