Add Isopycnal Curves to TS Plot

Adds isopycnal lines to an existing temperature-salinity plot. This is called by plotTS(), and may be called by the user also, e.g. if an image plot is used to show TS data density.

drawIsopycnals(
  nlevels = 6,
  levels,
  rotate = TRUE,
  rho1000 = FALSE,
  digits = 2,
  eos = getOption("oceEOS", default = "gsw"),
  trimIsopycnals = TRUE,
  cex = 0.75 * par("cex"),
  col = "darkgray",
  lwd = par("lwd"),
  lty = par("lty")
)

Arguments

nlevels

suggested number of density levels (i.e. isopycnal curves); ignored if levels is supplied.

levels

optional density levels to draw.

rotate

boolean, set to TRUE to write all density labels horizontally.

rho1000

boolean, set to TRUE to write isopycnal labels as e.g. 1024 instead of 24.

digits

number of decimal digits to use in label (supplied to round()).

eos

equation of state to be used, either "unesco" or "gsw".

trimIsopycnals

logical value (TRUE by default) that indicates whether to trim isopycnal curves (if drawn) to the region of temperature-salinity space for which density computations are considered to be valid in the context of the chosen eos; see the “Details” of the documentation for plotTS().

cex

size for labels.

col

color for lines and labels.

lwd

line width for isopycnal curves

lty

line type for isopycnal curves

Value

None.

References

Fofonoff, N. P., and R. C. Millard. "Algorithms for Computation of Fundamental Properties of Seawater." UNESCO Technical Papers in Marine Research. SCOR working group on Evaluation of CTD data; UNESCO/ICES/SCOR/IAPSO Joint Panel on Oceanographic Tables and Standards, 1983. https://unesdoc.unesco.org/ark:/48223/pf0000059832.

McDougall, Trevor J., David R. Jackett, Daniel G. Wright, and Rainer Feistel. "Accurate and Computationally Efficient Algorithms for Potential Temperature and Density of Seawater." Journal of Atmospheric and Oceanic Technology 20, no. 5 (May 1, 2003): 730-41. https://journals.ametsoc.org/jtech/article/20/5/730/2543/Accurate-and-Computationally-Efficient-Algorithms.

See also

Author

Dan Kelley