Julian-Day number to Julian century

Convert a Julian-Day number to a time in julian centuries since noon on January 1, 1900. The method follows reference 1 equation 15.1. The example reproduces the example provided by reference 1 example 15.a, with fractional error 3e-8.

julianCenturyAnomaly(jd)

Arguments

jd
a julian day number, e.g. as given by julianDay().

Value

Julian century since noon on January 1, 1900.

References

Meeus, Jean. Astronomical Formulas for Calculators. Second Edition. Richmond, Virginia, USA: Willmann-Bell, 1982.

See also

Author

Dan Kelley

Examples

t <- ISOdatetime(1978, 11, 13, 4, 35, 0, tz="UTC")
jca <- julianCenturyAnomaly(julianDay(t))
cat(format(t), "is Julian Century anomaly", format(jca, digits=8), "\n")

#> 1978-11-13 04:35:00 is Julian Century anomaly 0.78865684