Deprecated and Defunct Elements of the oce package

Certain functions and function arguments are still provided for compatibility with older versions of oce, but will be removed soon. The oce scheme for removing functions is similar to that used by Bioconductor: items are marked as "deprecated" in one release, marked as "defunct" in the next, and removed in the next after that. This goal is to provide a gentle migration path for users who keep their packages reasonably up-to-date.

Details

The following are marked "deprecated" in the present CRAN release of oce. Please use the replacement functions as listed below. The upcoming CRAN release of oce will mark these as "defunct", which is the last step before outright removal.

Deprecated	Replacement	Deprecated	Defunct	Removed

renameData()	oceRenameData()	1.1-2	1.1-3	1.1-4

The following are marked "defunct", so calling them in the the present version produces an error message that hints at a replacement function. Once a function is marked "defunct" on one CRAN release, it will be slated for outright deletion in a subsequent release.

Defunct	Replacement	Version

renameData()	oceRenameData()	1.5

byteToBinary()	rawToBits()	1.5

The following functions were removed after having been marked as "deprecated" in at least one CRAN release, and possibly as "defunct" in at least one CRAN release. (The version number in the table is the first version to lack the named function.)

Function	Replacement	Version

addColumn()	oceSetData()	1.1-2

ctdAddColumn()	oceSetData()	1.1-2

ctdUpdateHeader()	oceSetMetadata()	1.1-2

findInOrdered()	findInterval()	1.1-2

makeSection()	as.section()	0.9.24

mapMeridians()	mapGrid()	1.1-2

mapZones()	mapGrid()	1.1-2

oce.as.POSIXlt()	lubridate::parse_date_time()	1.1-2

Several oce function arguments are considered "deprecated", which means they will be marked "defunct" in the next CRAN release. These are normally listed in the help page for the function in question. A few that may be of general interest are also listed below.

The adorn argument was still being checked for (in the dots argument) until 2020 August 11.

The eos argument of swN2() was removed on 2019 April 11; for details, see the “Deprecation Notation” section of the documentation for swN2().

The endian argument of byteToBinary() will be removed sometime in the year 2017, and should be set to "big" in the meantime.

The parameters argument of plot,ctd-method() was deprecated on 2016-12-30. It was once used by plot,coastline-method() but has been ignored by that function since February 2016.

The orientation argument of plot,ctd-method() was deprecated on 2016-12-30. It was once used by plot,coastline-method() but has been ignored by that function since February 2016.

Several ‘oce’ function arguments are considered "defunct", which means they will be removed in the next CRAN release. They are as follows.

The fill argument of mapPlot() was confusing to users, so it was designated as deprecated in June 2016. (The confusion stemmed from subtle differences between plot() and polygon(), and the problem is that mapPlot() can use either of these functions, according to whether coastlines are to be filled.) The functionality is preserved, in the col argument.

See also