Plot a landsat Object

Plot the data within a landsat image, or information computed from the data. The second category includes possibilities such as an estimate of surface temperature and the "terralook" estimate of a natural-color view.

# S4 method for landsat
plot(
  x,
  band,
  which = 1,
  decimate = TRUE,
  zlim,
  utm = FALSE,
  col = oce.colorsPalette,
  drawPalette = TRUE,
  showBandName = TRUE,
  alpha.f = 1,
  red.f = 1.7,
  green.f = 1.5,
  blue.f = 6,
  offset = c(0, -0.05, -0.2, 0),
  transform = diag(c(red.f, green.f, blue.f, alpha.f)),
  debug = getOption("oceDebug"),
  ...
)

Arguments

x

a landsat object.

band

If given, the name of the band. For Landsat-8 data, this may be one of: "aerosol", "blue", "green", "red", "nir", "swir1", "swir2", "panchromatic", "cirrus", "tirs1", or "tirs2". For Landsat-7 data, this may be one of "blue", "green", "red", "nir", "swir1", "tirs1", "tirs2", "swir2", or "panchromatic". For Landsat data prior to Landsat-7, this may be one of "blue", "green", "red", "nir", "swir1", "tirs1", "tirs2", or "swir2". If band is not given, the ("tirs1") will be used if it exists in the object data, or otherwise the first band will be used. In addition to the above, using band="temperature" will plot an estimate of at-satellite brightness temperature, computed from the tirs1 band, and band="terralook" will plot a sort of natural color by combining the red, green, blue and nir bands according to the formula provided at https://lta.cr.usgs.gov/terralook/what_is_terralook (a website that worked once, but failed as of Feb 2, 2017).

which

Desired plot type; 1=image, 2=histogram.

decimate

An indication of the desired decimation, passed to imagep() for image plots. The default yields faster plotting. Some decimation is sensible for full-size images, since no graphical displays can show 16 thousand pixels on a side.

zlim

Either a pair of numbers giving the limits for the colorscale, or "histogram" to have a flattened histogram (i.e. to maximally increase contrast throughout the domain.) If not given, the 1 and 99 percent quantiles are calculated and used as limits.

utm

A logical value indicating whether to use UTS (easting and northing) instead of longitude and latitude on plot.

col

Either a function yielding colors, taking a single integer argument with the desired number of colors, or the string "natural", which combines the information in the red, green and blue bands and produces a natural-hue image. In the latter case, the band designation is ignored, and the object must contain the three color bands.

drawPalette

Indication of the type of palette to draw, if any. See imagep() for details.

showBandName

A logical indicating whether the band name is to plotted in the top margin, near the right-hand side.

alpha.f

Argument used if col="natural", to adjust colors with adjustcolor().

red.f

Argument used if col="natural", to adjust colors with adjustcolor(). Higher values of red.f cause red hues to be emphasized (e.g. dry land).

green.f

Argument used if col="natural", to adjust colors with adjustcolor(). Higher values of green.f emphasize green hues (e.g. forests).

blue.f

Argument used if band="terralook", to adjust colors with adjustcolor(). Higher values of blue.f emphasize blue hues (e.g. ocean).

offset

Argument used if band="terralook", to adjust colors with adjustcolor().

transform

Argument used if band="terralook", to adjust colors with adjustcolor().

debug

Set to a positive value to get debugging information during processing.

...

optional arguments passed to plotting functions.

Details

For Landsat-8 data, the band may be one of: "aerosol", "blue", "green", "red", "nir", "swir1", "swir2", "panchromatic", "cirrus", "tirs1", or "tirs2".

For Landsat-7 data, band may be one of "blue", "green", "red", "nir", "swir1", "tirs1", "tirs2", "swir2", or "panchromatic".

For Landsat data prior to Landsat-7, band may be one of "blue", "green", "red", "nir", "swir1", "tirs1", "tirs2", or "swir2".

If band is not given, the ("tirs1") will be used if it exists in the object data, or otherwise the first band will be used.

In addition to the above there are also some pseudo-bands that can be plotted, as follows.

Setting band="temperature" will plot an estimate of at-satellite brightness temperature, computed from the tirs1 band.

Setting band="terralook" will plot a sort of natural color by combining the red, green, blue and nir bands according to the formula provided at https://lta.cr.usgs.gov/terralook/what_is_terralook (a website that worked once, but failed as of Feb 2, 2017), namely that the red-band data are provided as the red argument of the rgb() function, while the green argument is computed as 2/3 of the green-band data plus 1/3 of the nir-band data, and the blue argument is computed as 2/3 of the green-band data minus 1/3 of the nir-band data. (This is not a typo: the blue band is not used.)

See also

Author

Dan Kelley