Plot an odf Object

Plot data contained within an ODF object, using oce.plot.ts() to create panels of time-series plots for all the columns contained in the odf object (or just those that contain at least one finite value, if blanks is FALSE). If the object's data slot does not contain time, then pairs() is used to plot all the elements in the data slot that contain at least one finite value. These actions are both crude and there are no arguments to control the behaviour, but this function is really just a stop-gap measure, since in practical work odf objects are usually cast to other types, and those types tend to have more useful plots.

# S4 method for odf
plot(x, blanks = TRUE, debug = getOption("oceDebug"))

Arguments

x

an odf object.

blanks

A logical value that indicates whether to include dummy plots for data items that lack any finite values.

debug

an integer specifying whether debugging information is to be printed during the processing. This is a general parameter that is used by many oce functions. Generally, setting debug=0 turns off the printing, while higher values suggest that more information be printed. If one function calls another, it usually reduces the value of debug first, so that a user can often obtain deeper debugging by specifying higher debug values.

See also

Author

Dan Kelley