Read a Sontek ADP File

Read a Sontek acoustic-Doppler profiler file (see reference 1).

read.adp.sontek(
  file,
  from = 1,
  to,
  by = 1,
  tz = getOption("oceTz"),
  longitude = NA,
  latitude = NA,
  type = c("adp", "pcadp"),
  monitor = FALSE,
  despike = FALSE,
  processingLog,
  debug = getOption("oceDebug"),
  ...
)

Arguments

file
a connection or a character string giving the name of the file to load. (For read.adp.sontek.serial, this is generally a list of files, which will be concatenated.)

from
indication of the first profile to read. This can be an integer, the sequence number of the first profile to read, or a POSIXt time before which profiles should be skipped, or a character string that converts to a POSIXt time (assuming UTC timezone). See “Examples”, and make careful note of the use of the tz argument. If from is not supplied, it defaults to 1.

to
an optional indication of the last profile to read, in a format as described for from. As a special case, to=0 means to read the file to the end. If to is not supplied, then it defaults to 0.

by
an optional indication of the stride length to use while walking through the file. If this is an integer, then by-1 profiles are skipped between each pair of profiles that is read, e.g. the default by=1 means to read all the data. (For RDI files only, there are some extra features to avoid running out of memory; see “Memory considerations”.)

tz
character string indicating time zone to be assumed in the data.

longitude
optional signed number indicating the longitude in degrees East.

latitude
optional signed number indicating the latitude in degrees North.

type
A character string indicating the type of instrument.

monitor
boolean value indicating whether to indicate the progress of reading the file, by using txtProgressBar() or otherwise. The value of monitor is changed to FALSE automatically, for non-interactive sessions.

despike
if TRUE, despike() will be used to clean anomalous spikes in heading, etc.

processingLog
if provided, the action item to be stored in the log. (Typically only provided for internal calls; the default that it provides is better for normal calls by a user.)

debug
a flag that turns on debugging. Set to 1 to get a moderate amount of debugging information, or to 2 to get more.

...
optional additional arguments that some (but not all) read.adp.*() functions pass to lower-level functions.

Value

An adp object. The contents of that object make sense for the particular instrument type under study, e.g. if the data file contains NMEA strings, then navigational data will be stored in an item called nmea in the data slot).

References

Information about Sontek profilers is available at https://www.sontek.com.

See also

Author

Dan Kelley and Clark Richards