Read a G1SST file

Read a G1SST file in the netcdf format provided by the ERDDAP server (see reference 1).

read.g1sst(file)

Arguments

file

character value containing the name of a netcdf file containing G1SST data.

Value

A g1sst object.

Details

As noted in the documentation for the g1sst class, one must be aware of the incorporation of model simulations in the g1sst product. For example, the code presented below might lead one to believe that the mapped field represents observations, whereas in fact it can be verified by consulting reference 2 (clicking and unclicking the radio button to show just the data) that the field mostly derives from simulation.

References

ERDDAP Portal https://coastwatch.pfeg.noaa.gov/erddap/

JPO OurOcean Portal https://ourocean.jpl.nasa.gov/SST/ (link worked in 2016 but was seen to fail 2017 Feb 2).

See also

Author

Dan Kelley

Examples

if (FALSE) {
# Construct query, making it easier to understand and modify.
day <- "2016-01-02"
lon0 <- -66.5
lon1 <- -64.0
lat0 <- 44
lat1 <- 46
source <- paste("https://coastwatch.pfeg.noaa.gov/erddap/griddap/",
                "jplG1SST.nc?",
                "SST%5B(", day, "T12:00:00Z)",
                "%5D%5B(", lat0, "):(", lat1, ")",
                "%5D%5B(", lon0, "):(", lon1, ")",
                "%5D", sep="")
if (!length(list.files(pattern="^a.nc$")))
    download.file(source, "a.nc")
d <- read.g1sst("a.nc")
plot(d, "SST", col=oceColorsTemperature)
if (requireNamespace("ocedata", quietly=TRUE)) {
    data(coastlineWorldFine, package="ocedata")
    lines(coastlineWorldFine[['longitude']],coastlineWorldFine[['latitude']])
}
}