Extract Something From a ctd Object

The [[ method works for all oce objects, i.e. objects inheriting from oce. The purpose is to insulate users from the internal details of oce objects, by looking for items within the various storage slots of the object. Items that are not actually stored in the object can also be extracted, including derived data such as potential temperature, the units of measurement for the data, data-quality flags, etc.

# S4 method for ctd
[[(x, i, j, ...)

Arguments

x
a ctd object.

i
Character string indicating the name of an item to extract.

j
Optional additional information on the i item.

...
Optional additional information (ignored).

Details

A two-step process is used to try to find the requested information. First, a class-specific function tries to find it, but if that fails, then a general function is used (see next section).

Details of the general method

If the specialized method produces no matches, the following generalized method is applied. As with the specialized method, the procedure hinges first on the values of i and, optionally, j. The work proceeds in steps, by testing a sequence of possible conditions in sequence.

A check is made as to whether i names one of the standard oce slots. If so, [[ returns the slot contents of that slot. Thus, x[["metadata"]] will retrieve the metadata slot, while x[["data"]] and x[["processingLog"]] return those slots.

If i is a string ending in the "Unit", then the characters preceding that string are taken to be the name of an item in the data object, and a list containing the unit is returned. This list consists of an item named unit, which is an expression(), and an item named scale, which is a string describing the measurement scale. If the string ends in " unit", e.g. x[["temperature unit"]] (note the space), then just the expression is returned, and if it ends in " scale", then just the scale is returned.

If i is a string ending in "Flag", then the corresponding data-quality flag is returned (or NULL if there is no such flag). For example, x[["salinityFlag"]] returns a vector of salinity flags if x is a ctd object.

If i is "sigmaTheta", then the value of swSigmaTheta() is called with x as the sole argument, and the results are returned. Similarly, swSigma0() is used if i="sigma0", and swSpice() is used if i="spice". Of course, these actions only make sense for objects that contain the relevant items within their data slot.

After these possibilities are eliminated, the action depends on whether j has been provided. If j is not provided, or is the string "", then i is sought in the metadata slot, and then in the data slot, returning whichever is found first. In other words, if j is not provided, the metadata slot takes preference over the data slot. However, if j is provided, then it must be either the string "metadata" or "data", and it directs where to look.

If none of the above-listed conditions holds, then NULL is returned, without the issuance of a warning or error message. (This silent operation is employed so that [[ will behave like the normal R version.)

Details of the specialized ctd method

Some uses of [[,ctd-method involve direct retrieval of items within the data slot of the ctd object, while other uses involve calculations based on items in that data slot. For an example, all ctd objects should hold an item named temperature in the data slot, so for example x[["temperature"]] will retrieve that item. By contrast, x[["sigmaTheta"]] is taken to be a request to compute \(\sigma_\theta\), and so it yields a call to swTheta(x) even if the data slot of x might happen to contain an item named theta. This can be confusing at first, but it tends to lead to fewer surprises in everyday work, for otherwise the user would be forced to check the contents of any ctd object under analysis, to determine whether that item will be looked up or computed. Nothing is lost in this scheme, since the data within the object are always accessible with oceGetData().

It should be noted that the accessor is set up to retrieve quantities in conventional units. For example, read.ctd.sbe() is used on a .cnv file that stores pressure in psi, it will be stored in the same unit within the ctd object, but x[["pressure"]] will return a value that has been converted to decibars. (To get pressure in PSI, use x[["pressurePSI"]].) Similarly, temperature is returned in the ITS-90 scale, with a conversion having been performed with T90fromT68(), if the object holds temperature in IPTS-68. Again, temperature on the IPTS-68 scale is returned with x@data$temperature.

This preference for computed over stored quantities is accomplished by first checking for computed quantities, and then falling back to the general [[ method if no match is found.

Some quantities are optionally computed. For example, some data files (e.g. the one upon which the section() dataset is based) store nitrite along with the sum of nitrite and nitrate, the latter with name NO2+NO3. In this case, e.g. x[["nitrate"]] will detect the setup, and subtract nitrite from the sum to yield nitrate.

The list given below provides notes on some quantities that are, or may be, computed.

conductivity without a second argument (e.g. a[["conductivity"]]) returns the value stored in the object. However, if a second argument is given, and it is string specifying a unit, then conversion is made to that unit. The permitted units are: either "" or "ratio" (for ratio), "uS/cm", "mS/cm" and "S/m". The calculations are based on the definition of conductivity ratio as the ratio between measured conductivity and the standard value 4.2914 S/m.

CT or Conservative Temperature: Conservative Temperature, computed with gsw::gsw_CT_from_t().

density: seawater density, computed with swRho(x). (Note that it may be better to call that function directly, to gain control of the choice of equation of state, etc.)

depth: Depth in metres below the surface, computed with swDepth(x).

N2: Square of Brunt-Vaisala frequency, computed with swN2(x).

potential temperature: Potential temperature in the UNESCO formulation, computed with swTheta(x). This is a synonym for theta.

Rrho: Density ratio, computed with swRrho(x).

SA or Absolute Salinity: Absolute Salinity, computed with gsw::gsw_SA_from_SP(). The calculation involves location as well as measured water properties. If the object x does not containing information on the location, then 30N and 60W is used for the calculation, and a warning is generated.

sigmaTheta: A form of potential density anomaly, computed with swSigmaTheta(x).

sigma0 Equal to sigmaTheta, i.e. potential density anomaly referenced to a pressure of 0dbar, computed with swSigma0(x).

sigma1: Potential density anomaly referenced to a pressure of 1000dbar, computed with swSigma1(x).

sigma2: Potential density anomaly referenced to a pressure of 2000dbar, computed with swSigma2(x).

sigma3: Potential density anomaly referenced to a pressure of 3000dbar, computed with swSigma3(x).

sigma4: potential density anomaly referenced to a pressure of 4000dbar, computed with swSigma4(x).

SP: Salinity on the Practical Salinity Scale, which is salinity in the data slot.

spice or spiciness0: a variable that is in some sense orthogonal to density, calculated with swSpice(x). Note that this is defined differently for eos="unesco" and eos="gsw".

SR: Reference Salinity computed with gsw::gsw_SR_from_SP().

Sstar: Preformed Salinity computed with gsw::gsw_SR_from_SP(). See SA for a note on longitude and latitude.

theta: potential temperature in the UNESCO formulation, computed with swTheta(x). This is a synonym for potential temperature.

time: returns either a vector of times, a single time, or NULL. A vector is returned if time is present in the data slot, or if a time can be inferred from other entries in the data slot (some of which, such as the common timeS, also employ startTime within the metadata slot). A single value is returned if the dataset only has information on the start time (which is stored as startTime within the metadata slot. If it is impossible to determine the sampling time, then NULL is returned. These time variants occur, in the present version of oce, only for data read by read.ctd.sbe(), the documentation of which explains how times are computed.

z: Vertical coordinate in metres above the surface, computed with swZ(x).

See also

Author

Dan Kelley

Examples

data(ctd)
head(ctd[["temperature"]])

#> [1] 14.22109 14.22649 14.22509 14.22219 14.22669 14.23318