Summarize an Argo Object

Summarizes some of the data in an argo object.

# S4 method for argo
summary(object, ...)

Arguments

...

Further arguments passed to or from other methods.

object

an

object of class "argo", usually, a result of a call to read.argo().

Value

A matrix containing statistics of the elements of the data slot.

Details

Pertinent summary information is presented.

See also

Author

Dan Kelley

Examples

library(oce)
data(argo)
summary(argo)
#> Argo Summary
#> ------------
#> 
#> * Source:              "/Users/kelley/git/oce/create_data/argo/6900388_prof.nc"
#> * id:                  "6900388"
#> * feature type:        "trajectoryProfile"
#> * Profiles:            210 delayed; 0 adjusted; 13 realtime
#> * Time ranges from 2005-10-29 13:57:42 to 2011-11-27 17:58:39 with 223 samples and mean increment 10.00075 day
#> * Data Overview
#> 
#>                                             Min.    Mean    Max.   Dim. NAs        OriginalName
#>     latitude [°N]                         48.743  56.715  64.335    223   0            LATITUDE
#>     longitude [°E]                        -60.52 -37.077 -21.385    223   0           LONGITUDE
#>     pressure [dbar]                          3.5  521.31  6534.6 56x223 106                PRES
#>     pressureAdjusted [dbar]                  3.6  520.34  1778.9 56x223 874       PRES_ADJUSTED
#>     pressureAdjustedError [dbar]             2.4     2.4     2.4 56x223 874 PRES_ADJUSTED_ERROR
#>     salinity [PSS-78]                          0  34.911  47.899 56x223 106                PSAL
#>     salinityAdjusted [PSS-78]              32.85  34.935  35.842 56x223 840       PSAL_ADJUSTED
#>     salinityAdjustedError [PSS-78]          0.01    0.01    0.01 56x223 840 PSAL_ADJUSTED_ERROR
#>     temperature [°C, ITS-90]              -1.564  6.1216  55.997 56x223 106                TEMP
#>     temperatureAdjusted [°C, ITS-90]      -1.564  6.0916  14.914 56x223 837       TEMP_ADJUSTED
#>     temperatureAdjustedError [°C, ITS-90]  0.002   0.002   0.002 56x223 837 TEMP_ADJUSTED_ERROR
#> 
#> * Data-quality Flag Scheme
#> 
#>     name    "argo"
#>     mapping list(not_assessed=0, passed_all_tests=1, probably_good=2, probably_bad=3, bad=4, changed=5, not_used_6=6, not_used_7=7, estimated=8, missing=9)
#>     default c(0, 3, 4, 9)
#> 
#> * Data-quality Flags
#> 
#>     pressure:            "1" 12327, "4" 55
#>     pressureAdjusted:    "2" 11614
#>     salinity:            "1" 12348, "4" 34
#>     salinityAdjusted:    "2" 11648
#>     temperature:         "1" 12362, "4" 20
#>     temperatureAdjusted: "2" 11651
#> 
#> * Processing Log
#> 
#>     - 2020-11-09 16:05:39 UTC: `create 'argo' object`
#>     - 2020-11-09 16:05:39 UTC: `initializeFlagScheme(object, name="argo", mapping=list(not_assessed=0,passed_all_tests=1,probably_good=2,probably_bad=3,bad=4,changed=5,not_used_6=6,not_used_7=7,estimated=8,missing=9)), default=c(0,3,4,9))`
#>     - 2020-11-09 16:05:39 UTC: `read.argo(file="/Users/kelley/git/oce/create_data/argo/6900388_prof.nc")`