Summarize a CTD Object

Summarizes some of the data in a ctd object, presenting such information as the station name, sampling location, data ranges, etc. If the object was read from a .cnv file or a .rsk file, then the OriginalName column for the data summary will contain the original names of data within the source file.

# S4 method for ctd
summary(object, ...)

Arguments

object
a ctd object.

...
Further arguments passed to or from other methods.

See also

Examples

library(oce)
data(ctd)
summary(ctd)

#> CTD Summary
#> -----------
#> 
#> * Instrument:          SBE 25 
#> * Temp. serial no.:    1140
#> * Cond. serial no.:    832
#> * File:                "/Users/kelley/git/oce/create_data/ctd/ctd.cnv"
#> * Original file:       c:\seasoft3\basin\bed0302.hex
#> * Start time:          2003-10-15 15:38:38
#> * Sample interval:     1 s
#> * Cruise:              Halifax Harbour
#> * Vessel:              Divcom3
#> * Station:             Stn 2
#> * Location:            44.684N63.644W
#> * Data Overview
#> 
#>                                 Min.   Mean   Max. Dim. NAs OriginalName
#>     scan                         130    220    310  181   0         scan
#>     timeS [s]                    129    219    309  181   0        timeS
#>     pressure [dbar]             1.48 22.885 44.141  181   0           pr
#>     depth [m]                  1.468 22.698 43.778  181   0         depS
#>     temperature [°C, IPTS-68]  2.919 7.5063 14.237  181   0         t068
#>     salinity [PSS-78]         29.916 31.219 31.498  181   0        sal00
#>     flag                           0      0      0  181   0         flag
#> 
#> * Processing Log
#> 
#>     - 2018-11-14 20:03:47 UTC: `create 'ctd' object`
#>     - 2018-11-14 20:03:47 UTC: `read.ctd.sbe(file = file, debug = 10, processingLog = processingLog)`
#>     - 2018-11-14 20:03:47 UTC: `oce.edit(x = ctd, item = "startTime", value = as.POSIXct(gsub("1903",     "2003", format(ctd[["startTime"]])), tz = "UTC") + 4 * 3600,     reason = "file had year=1903, instead of 2003", person = "Dan Kelley")`