Water depth

Compute depth below the surface (i.e. a positive number within the water column) based on pressure and latitude. (Use swZ() to get the vertical coordinate, which is negative within the water column.)

swDepth(pressure, latitude = 45, eos = getOption("oceEOS", default = "gsw"))

Arguments

pressure

either pressure (dbar), in which case lat must also be given, or a ctd object, in which case lat will be inferred from the object.

latitude

Latitude in \(^\circ\)N or radians north of the equator.

eos

indication of formulation to be used, either "unesco" or "gsw".

Value

Depth below the ocean surface, in metres.

Details

If eos="unesco" then depth is calculated from pressure using Saunders and Fofonoff's method, with the formula refitted for 1980 UNESCO equation of state (reference 1). If eos="gsw", then gsw::gsw_z_from_p() from the gsw package (references 2 and 3) is used.

References

Unesco 1983. Algorithms for computation of fundamental properties of seawater, 1983. Unesco Tech. Pap. in Mar. Sci., No. 44, 53 pp.

IOC, SCOR, and IAPSO (2010). The international thermodynamic equation of seawater-2010: Calculation and use of thermodynamic properties. Technical Report 56, Intergovernmental Oceanographic Commission, Manuals and Guide.

McDougall, T.J. and P.M. Barker, 2011: Getting started with TEOS-10 and the Gibbs Seawater (GSW) Oceanographic Toolbox, 28pp., SCOR/IAPSO WG127, ISBN 978-0-646-55621-5.

See also

Author

Dan Kelley

Examples

d <- swDepth(10, 45)