Show some values from a vector

This is similar to str(), but it shows data at the first and last of the vector, which can be quite helpful in debugging.

vectorShow(v, msg = "", postscript = "", digits = 5, n = 2L)

Arguments

v
the vector.

msg
optional character value indicating a message to show, introducing the vector. If not provided, then a message is created from v. If msg is a non-empty string, then that string is pasted together with a colon (unless msg already contains a colon), before pasting a summary of data values.

postscript
optional character value indicating an optional message to append at the end of the return value.

digits
for numerical values of v, this is the number of digits to use, in formatting the numbers with format(); otherwise, digits is ignored.

n
number of elements to show at start and end. If n is negative, then all the elements are shown.

Value

A string ending in a newline character, suitable for display with cat() or oceDebug().

Author

Dan Kelley

Examples

vectorShow(pi)

#> [1] "pi: 3.1416\n"

vectorShow(volcano)

#> [1] "volcano [1:87, 1:61]: 100, 101, ..., 94, 94\n"

knot2mps <- 0.5144444
vectorShow(knot2mps, postscript="knots per m/s")

#> [1] "knot2mps: 0.51444 knots per m/s\n"

vectorShow("January", msg="The first month is")

#> [1] "The first month is: January\n"