Convert ADP2CP adp object From XYZ to ENU Coordinates

This function will b in active development through the early months of 2019, and both the methodology and user interface may change without notice. Only developers (or invitees) should be trying to use this function.

xyzToEnuAdpAD2CP(x, declination = 0, debug = getOption("oceDebug"))

Arguments

x

an adp object created by read.adp.ad2cp().

declination

IGNORED at present, but will be used at some later time.

debug

an integer specifying whether debugging information is to be printed during the processing. This is a general parameter that is used by many oce functions. Generally, setting debug=0 turns off the printing, while higher values suggest that more information be printed. If one function calls another, it usually reduces the value of debug first, so that a user can often obtain deeper debugging by specifying higher debug values.

Value

An object with data$v[,,1:3] altered appropriately, and x[["oceCoordinate"]] changed from xyz to enu.

Limitations

This only works if the instrument orientation is "AHRS", and even that is not tested yet. Plus, as noted, the declination is ignored.

References

Nortek AS. “Signature Integration 55|250|500|1000kHz.” Nortek AS, 2017.

Nortek AS. “Signature Integration 55|250|500|1000kHz.” Nortek AS, 2018. https://www.nortekgroup.com/assets/software/N3015-007-Integrators-Guide-AD2CP_1018.pdf.

See also

Author

Dan Kelley